AWCPTF {GSIF}	R Documentation

Available soil water capacity

Description

Derive available soil water capacity (in cubic-meter per cubic-meter) based on a Pedo-Transfer Function developed using the Africa Soil Profile Database (Wösten et al. 2013).

Usage

AWCPTF(SNDPPT, SLTPPT, CLYPPT, ORCDRC, 
    BLD=1682, CEC, PHIHOX, h1=-10, h2=-20, h3=-31.6, 
    pwp=-1585, PTF.coef, fix.values=TRUE )

Arguments

SNDPPT
numeric; sand content in percent

SLTPPT
numeric; silt content in percent

CLYPPT
numeric; clay content in percent

ORCDRC
numeric; soil organic carbon concentration in permille or g / kg

BLD
numeric; bulk density in kg / cubic-meter for the horizon/solum

CEC
numeric; Cation Exchange Capacity in cmol per kilogram

PHIHOX
numeric; soil pH in water suspension

h1
numeric; moisture potential in kPa e.g. -10 (pF 2.0)

h2
numeric; moisture potential in kPa e.g. -20 (pF 2.3)

h3
numeric; moisture potential in kPa e.g. -31.6 (pF 2.5)

pwp
numeric; moisture potential at wilting point in kPa e.g. -1585 (pF 4.2)

PTF.coef
data.frame; optional conversion coefficients (Pedo-Transfer Function)

fix.values
logical; specifies whether to correct values of textures and bulk density to avoid creating non-sensical values

Value

Returns a data frame with the following columns:

AWCh1: available soil water capacity (volumetric fraction) for h1;

AWCh2: available soil water capacity (volumetric fraction) for h2;

AWCh3: available soil water capacity (volumetric fraction) for h3;

WWP: available soil water capacity (volumetric fraction) until wilting point;

Note

Pedotransfer coefficients (PTF.coef) developed by Hodnett and Tomasella (2002). fix.values will correct sand, silt and clay fractions so they sum up to 100, and will replace bulk density values using global minimum maximum values.

Author(s)

Johan Leenaars and Maria Ruiperez Gonzalez

References

Hodnett, M. G., & Tomasella, J. (2002). Marked differences between van Genuchten soil water-retention parameters for temperate and tropical soils: a new water-retention pedo-transfer functions developed for tropical soils. Geoderma, 108(3), 155-180.

Wösten, J. H. M., Verzandvoort, S. J. E., Leenaars, J. G. B., Hoogland, T., & Wesseling, J. G. (2013). Soil hydraulic information for river basin studies in semi-arid regions. Geoderma, 195, 79-86.

Examples


SNDPPT = 30 
SLTPPT = 25 
CLYPPT = 48 
ORCDRC = 23 
BLD = 1200 
CEC = 12 
PHIOX = 6.4
x <- AWCPTF(SNDPPT, SLTPPT, CLYPPT, ORCDRC, BLD, CEC, PHIHOX)
x
## predict AWC for the whole AfSP DB:
data(afsp)
names(afsp$horizons) 
BLDf <- ifelse(is.na(afsp$horizons$BLD), 
   mean(afsp$horizons$BLD, na.rm=TRUE), afsp$horizons$BLD)
xd <- AWCPTF(afsp$horizons$SNDPPT, afsp$horizons$SLTPPT, 
  afsp$horizons$CLYPPT, afsp$horizons$ORCDRC, 
  BLD=BLDf*1000, afsp$horizons$CEC, afsp$horizons$PHIHOX)
str(xd)

[Package GSIF version 0.4-2 Index]