ERDICM {GSIF}	R Documentation

Effective Rooting Zone depth

Description

Derive Effective Rooting Zone depth i.e. an effective depth suitable for plant growth. Usually minimum depth of soil out of three standard rooting depths: limiting soil properties, depth to water-stagnating layer and depth to bedrock.

Usage

ERDICM(UHDICM, LHDICM, minimum.LRI, DRAINFAO, BDRICM, 
    threshold.LRI=20, srd=150, drain.depths, smooth.LRI=TRUE)

Arguments

UHDICM
numeric; upper horizon depth in cm

LHDICM
numeric; lower horizon depth in cm

minimum.LRI
numeric; minimum Limiting Rootability index

DRAINFAO
factor; FAO drainage class e.g. "V", "P", "I", "M", "W", "S", "E"

BDRICM
numeric; depth to bedrock in cm

threshold.LRI
numeric; treshold index for LRI

srd
numeric; maximum depth of interest

drain.depths
data.frame; estimate effective rooting depth per drainage class (DRAINFAO)

smooth.LRI
logical; specify whether to smooth LRI values using splines

Value

Returns a vector of effective rooting depth in cm.

Author(s)

Johan Leenaars, Maria Ruiperez Gonzalez and Tomislav Hengl

See Also

LRI

[Package GSIF version 0.4-5 Index]