GSIF.env {GSIF}	R Documentation

GSIF specific environmental variables / paths

Description

Sets the environmental, package specific parameters and settings (URLs, names, default cell size and similar) that can be later on passed to other functions.

Usage

GSIF.env(wps.server = "http://wps.worldgrids.org", 
    ref_CRS = "+proj=longlat +datum=WGS84",
    NAflag = -99999, 
    license_url = "https://creativecommons.org/licenses/by-sa/4.0/", 
    project_url = "http://gsif.r-forge.r-project.org/", 
    sldepths = c(0, 5, 15, 30, 60, 100, 200),
    stdepths = c(-2.5, -10, -22.5, -45, -80, -150)/100, 
    stsize = c(5, 10, 15, 30, 40, 100)/100, 
    cellsize = rev(c(6/120, 3/120, 1/120, 1/240, 1/600, 1/1200, 1/3600)),
    REST.server = 'https://rest.soilgrids.org/',
    attributes = c("ORCDRC","PHIHOX","SNDPPT","SLTPPT","CLYPPT",
                   "CFRVOL","CECSOL","BLDFIE","TAXNWRB","TAXOUSDA"),
    TimeSpan = list(begin=as.POSIXct("1950-01-01"), end=as.POSIXct("2005-12-30")),
    show.env = TRUE)

Arguments

wps.server
character; location of the WPS server

ref_CRS
the referent CRS proj4string ("+proj=longlat +datum=WGS84")

NAflag
the default missing value flag (usually "-99999")

license_url
the default license URL

project_url
the default location of the package documentation

sldepths
numeric; standard depths based on GlobalSoilMap.net specifications

stdepths
numeric; standard depths at the centre of layer

stsize
numeric; standard horizon thicknesses

cellsize
numeric; standard grid cell sizes on WGS84 geographical coordinates

REST.server
character; location of the SoilGrids REST service

attributes
character; default soil variables of interest

TimeSpan
list; default begin end times (temporal coverage of SoilGrids)

show.env
logical; specify whether to print all environmental parameters

Note

To further customize the GSIF options, consider putting:

library(GSIF); GSIF.env(..., show.env = FALSE)

in your "/etc/Rprofile.site".

Author(s)

Tomislav Hengl

Examples


# environmental variables:
GSIF.env()
get("cellsize", envir = GSIF.opts)

[Package GSIF version 0.5-3 Index]