GlobalSoilMap-class {GSIF}	R Documentation

A class for GlobalSoilMap soil property maps

Description

A class containing predictions of target soil property at six standard depths following the GlobalSoilMap.net specifications: sd1 = 2.5 cm (0–5), sd2 = 10 cm (5–15), sd3 = 22.5 cm (15–30), sd4 = 45 cm (30–60), sd5 = 80 cm (60–100), sd6 = 150 cm (100–200).

Slots

varname:

object of class "character"; abbreviated variable name registered in the Global Soil Data registry

TimeSpan:

object of class "list"; contains begin and end of the sampling period of class "POSIXct"

sd1:

object of class "SpatialPixelsDataFrame"; predictions and variances, or number of realizations of the target variable at depth 2.5 cm (0–5)

sd2:

object of class "SpatialPixelsDataFrame"; predictions and variances, or number of realizations of the target variable at depth 10 cm (5–15)

sd3:

object of class "SpatialPixelsDataFrame"; predictions and variances, or number of realizations of the target variable at depth 22.5 cm (15–30)

sd4:

object of class "SpatialPixelsDataFrame"; predictions and variances, or number of realizations of the target variable at depth 45 cm (30–60)

sd5:

object of class "SpatialPixelsDataFrame"; predictions and variances, or number of realizations of the target variable at depth 80 cm (60–100)

sd6:

object of class "SpatialPixelsDataFrame"; predictions and variances, or number of realizations of the target variable at depth 150 cm (100–200)

References

Hartemink, A. E., Hempel, J., Lagacherie, P., McBratney, A., McKenzie, N., MacMillan, R. A., ... & Zhang, G. L. (2010). GlobalSoilMap.net — A New Digital Soil Map of the World. In Digital Soil Mapping (pp. 423–428). Springer Netherlands.

Sanchez, P. A., S. Ahamed, F. Carre, A. E. Hartemink, J. Hempel, J. Huising, P. Lagacherie, A. B. McBratney, N. J. McKenzie, M L. de Mendonça-Santos, et al., (2009) Digital Soil Map of the World. Science, 325(5941): 680–681.

See Also

SoilGrids-class, SpatialComponents-class, geosamples-class

[Package GSIF version 0.4-2 Index]