GlobalSoilMap-class {GSIF}	R Documentation

A class for GlobalSoilMap soil property maps

Description

A class containing multiple realizations of soil properties following the GlobalSoilMap.net specifications, i.e. predictions of some of the target global soil properties at six depths (sd1 = 2.5 cm, sd2 = 10 cm, sd3 = 22.5 cm, sd4 = 45 cm, sd5 = 80 cm, sd6 = 150 cm).

Slots

varname:

object of class "character"; abbreviated variable name registered in the Global Soil Data registry

sd1:

object of class "SpatialPixelsDataFrame"; number of realizations of the target variable at depth 2.5 cm

sd2:

object of class "SpatialPixelsDataFrame"; number of realizations of the target variable at depth 10 cm

sd3:

object of class "SpatialPixelsDataFrame"; number of realizations of the target variable at depth 22.5 cm

sd4:

object of class "SpatialPixelsDataFrame"; number of realizations of the target variable at depth 45 cm

sd5:

object of class "SpatialPixelsDataFrame"; number of realizations of the target variable at depth 80 cm

sd6:

object of class "SpatialPixelsDataFrame"; number of realizations of the target variable at depth 150 cm

validation:

object of class "SpatialPointsDataFrame"; cross-validation data set with values of the target variable at six depths

Methods

summary

signature(x = "GlobalSoilMap"): generates summary statistics for the object

Author(s)

Tomislav Hengl and Robert A. MacMillan

See Also

SpatialComponents-class

Examples


## Not run: # load soil samples from the plotKML package: 
library(plotKML)
data(eberg)
# 30 percent subsample:
eberg <- eberg[runif(nrow(eberg)) < .3,]
# sites table:
s.lst <- c("ID", "soiltype", "TAXGRSC", "X", "Y")
sites <- eberg[,s.lst]
# get horizons table:
horizons <- getHorizons(eberg, idcol="ID", sel=c("UHDICM","LHDICM","SNDMHT","SLTMHT","CLYMHT"))
# create object of type "SoilProfileCollection"
eberg.spc <- join(horizons, sites, type='inner')
depths(eberg.spc) <- ID ~ UHDICM + LHDICM
site(eberg.spc) <- as.formula(paste("~", paste(s.lst[-1], collapse="+"), sep=""))
coordinates(eberg.spc) <- ~X+Y
proj4string(eberg.spc) <- CRS("+init=epsg:31467")
# convert to geosamples:
eberg.geo <- as.geosamples(eberg.spc)
# prepare covariates:
data(eberg_grid)
gridded(eberg_grid) <- ~x+y
proj4string(eberg_grid) <- CRS("+init=epsg:31467")
eberg_spc <- spc(eberg_grid, as.formula("~ PRMGEO6+DEMSRT6+TWISRT6+TIRAST6"))
# fit gstat Model:
formulaString = as.formula(paste("observedValue ~ ", paste(names(eberg_spc@predicted), collapse="+"), "+ ns(altitude, df=5)"))
SNDMHT.m <- fit.gstatModel(eberg.geo, formulaString, covariates=eberg_spc, methodid="SNDMHT")
summary(SNDMHT.m@regModel)

## End(Not run)

[Package GSIF version 0.2-1 Index]