NASIS {GSIF}	R Documentation

NASIS pedon data for Indiana State

Description

A subset of the National Soil Information System (NASIS) pedon data covering the Indiana State. Contains a total of 489 representative soil profiles.

Usage

data(NASIS)

Format

Sites table. Contains the following columns:

siteiid

factor; unique name that identifies a particular site

obsdate

POSIXct; sampling date

usiteid

factor; unique name set by the surveyor e.g. "92IN025001"

LON

numeric; longitude in decimal degrees (NAD83 datum)

LAT

numeric; latitude in decimal degrees (NAD83 datum)

pedonpurpose

factor; intended purpose of the profile description e.g. "full pedon description"

pedontype

factor; the description represents in relation to a series, component, etc e.g. "modal pedon for series"

geomposhill

factor; geomorphic component - hills e.g. "interfluve"

hillslopeprof

factor; two dimensional slope segments of a hillslope that have similar geometric, erosional, or depositional characteristics e.g. "footslope"

shapeacross

factor; two dimensional shape profile e.g. "concave"

shapedown

factor; the longitudinal shape of the slope e.g. "concave"

drainagecl

factor; the natural drainage conditions of the soil e.g. "very poorly"

siteperm

factor; site permeability e.g. "moderately slow"

runoff

factor; potential runoff class for the soil e.g. "very low"

soinmassamp

factor; soil name as sampled e.g. "Zanesville"

taxorder

factor; the Keys to Soil Taxonomy order class e.g. "alfisols"

taxsuborder

factor; the Keys to Soil Taxonomy suborder class e.g. "udalfs"

taxgrtgroup

factor; the Keys to Soil Taxonomy group class e.g. "fraglossudalfs"

taxsubgrp

factor; the Keys to Soil Taxonomy subgroup class e.g. "oxyaquic fraglossudalfs"

taxpartsize

factor; the Keys to Soil Taxonomy family / particle class e.g. "fine-silty"

taxtempregime

factor; the soil temperature regime e.g. "mesic"

Horizons table. Contains the following columns:

siteiid

factor; unique name that identifies a particular site

phiid

factor; unique name that identifies a particular horizon

hzname

factor; pedological horizon designation e.g. "Btx1"

seqnum

integer; horizon sequence number

hzdept

integer; horizon top depth in cm

hzdepb

integer; horizon bottom depth in cm

texture

factor; texture by hand class determined on the field e.g. "SIL"

rupresblkmst

integer; rupture moist class

phfield

numeric; pH as measured on the field

phdetermeth

factor; pH determination method

bounddistinct

integer; boundary distinctness grade

boundtopo

integer; horizontal shape of the interface between adjacent soil horizons

Note

The original NASIS database contains about 25 thousand profiles. Locations of samples were selected by experienced surveyors and are often clustered in some parts of the state (see plot below).
The variables listed above were selected to test various Digital Soil Mapping algorithms and represent only a subset of the NASIS data. Please refer to the NASIS technical documentation for more information about the specific columns.

Author(s)

The NASIS server is located at the USDA Kansas City data center. This subset has been prepared for the purpose of testing various DSM methods by Tomislav Hengl (tom.hengl@wur.nl). The processing steps are described in this script.

References

National Soil Information System (http://soils.usda.gov/technical/nasis/)

soilDB package (http://cran.r-project.org/package=soilDB)

See Also

NCSS, hoosier, geochm

Examples


library(aqp)
data(NASIS)
plot(NASIS$sites$LON, NASIS$sites$LAT)
aea.prj <- "+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=23 +lon_0=-96 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs"
NASIS.spc <- join(NASIS$horizons, NASIS$sites, type='inner')
depths(NASIS.spc) <- siteiid ~ hzdept + hzdepb
site(NASIS.spc) <- ~ LON + LAT + taxorder
# generate SpatialPoints:
coordinates(NASIS.spc) <- ~ LON + LAT
proj4string(NASIS.spc@sp) <- CRS("+proj=longlat +datum=NAD83")
# soil covariate layers:
data(indiana)
border <- SpatialLines(list(Lines(list(Line(indiana)), "border")))
proj4string(border) <- aea.prj
par(mar=c(0,0,0,0))
plot(spTransform(border, NASIS.spc@sp@proj4string))
points(NASIS.spc@sp, pch="+", cex=.7)

[Package GSIF version 0.1-1 Index]