SoilGrid.validator {GSIF}	R Documentation

Validate SoilGrid (spatial predictions)

Description

Validate SoilGrid (spatial predictions) i.e. soil property maps following the GSIF validation protocol.

Usage


SoilGrid.validator(obj, domain, ground.truth, N.sample=2000, 
   xml.file, z.lim, md.type="INSPIRE", test.URL=FALSE)

Arguments

obj
"GDALobj" object i.e. a pointer to a spatial layer of interest (single slice)

domain
"GDALobj" object i.e. a pointer to a spatial layer contain soil mask

ground.truth
"SpatialPointsDataFrame"; contains values of the target variable at exactly the same depth / same support size sampled either using Simple Random Sampling or regular sampling on a grid

N.sample
integer; random sampling size

xml.file
character; metadata file (should have the same name as obj file)

z.lim
numeric; upper and lower physical limits

md.type
character; metadata standard (currently INSPIRE)

test.URL
logical; specifies whether to validate XML schema / test download times and proj4 string

Value

Returns a data frame with validation results. Explanation of codes is available in the SoilGrids.org data validation protocol.

Note

One SoilGrid layer (2D slice) basically contains predictions on a regular grid for a specific soil depth (at either point or block support). The ground.truth data must refer to the exactly the same depth and the same support size and should ideally be collected using some probability spatial sampling (see e.g. sp::spsample). To estimate values of soil properties at standard depths, consider using mpspline function.
Numeric resolution is derived as estimated RMSE/2. Numeric resolution can be best specified as Attribute_Measurement_Resolution (the smallest unit increment to which an attribute value is measured).
Increasing N.sample can lead to more precise results at the cost of higher computing time.

Author(s)

Tomislav Hengl

See Also

plotKML::spMetadata

[Package GSIF version 0.4-5 Index]