SpatialComponents-class {GSIF}	R Documentation

A class for gridded components derived using the spc method

Description

A class containing a list of gridded components and results of principal component analysis.

Slots

predicted:

object of class "SpatialPixelsDataFrame"; predicted values for components

pca:

object of class "list"; output objects from the stats::prcomp process — contains objects: 'stdev', 'rotation', 'center' and 'scale'

Author(s)

Tomislav Hengl

See Also

spc

Examples

 
## Not run: # load soil samples from the plotKML package:
library(plotKML)
data(eberg)
# 20 percent subsample:
eberg <- eberg[runif(nrow(eberg)) < .2,]
# sites table:
s.lst <- c("ID", "soiltype", "TAXGRSC", "X", "Y")
sites <- eberg[,s.lst]
# get horizons table:
horizons <- getHorizons(eberg, idcol="ID", sel=c("UHDICM","LHDICM","SNDMHT","SLTMHT","CLYMHT"))
# create object of type "SoilProfileCollection"
eberg.spc <- join(horizons, sites, type='inner')
depths(eberg.spc) <- ID ~ UHDICM + LHDICM
site(eberg.spc) <- as.formula(paste("~", paste(s.lst[-1], collapse="+"), sep=""))
coordinates(eberg.spc) <- ~X+Y
proj4string(eberg.spc) <- CRS("+init=epsg:31467")
# convert to geosamples:
eberg.geo <- as.geosamples(eberg.spc)
# prepare covariates:
data(eberg_grid)
gridded(eberg_grid) <- ~x+y
proj4string(eberg_grid) <- CRS("+init=epsg:31467")
eberg_spc <- spc(eberg_grid, as.formula("~ PRMGEO6+DEMSRT6+TWISRT6+TIRAST6"))
# overlay with SPCs (create the regression matrix):
SNDMHT.eberg <- overlay(eberg_spc@predicted, eberg.geo, methodid="SNDMHT")
# fit a glm:
formulaString = observedValue~PC1+PC2+PC3+PC4+PC5+PC6+PC7+PC8+PC9+PC10+PC11+ns(altitude, df=4)
SNDMHT.glm <- glm(formulaString, SNDMHT.eberg, family=gaussian())
summary(SNDMHT.glm)

## End(Not run)

[Package GSIF version 0.2-2 Index]