SpatialMemberships-class {GSIF}	R Documentation

A class for membership maps derived using the fkmeans classification

Description

A class containing a list of gridded maps and results of model fitting.

Slots

predicted:

object of class "SpatialPixelsDataFrame"; predicted values (factor)

model:

object of class "multinom"; output object from the nnet::multinom method

mu:

object of class "SpatialPixelsDataFrame"; a list of predicted memberships

class.c:

object of class "matrix"; class centres

class.sd:

object of class "matrix"; class deviations

confusion:

object of class "matrix"; confusion matrix

Author(s)

Tomislav Hengl

See Also

spfkm, SpatialComponents-class

[Package GSIF version 0.2-5 Index]