USDA.TT.im {GSIF}	R Documentation

Probability density for texture triangle

Description

Probability density for texture triangle (USDA system) based on global soil profile data (http://soilprofiles.org).

Usage

data(USDA.TT.im)

Format

The USDA.TT.im data frame contains the following columns:

v

numeric; probability density derived using the soiltexture::TT.kde2d function and global soil profile data

TEXMHT

factor; USDA soil texture class estimated by hand (one of the following: "C", "SiC", "SC", "CL", "SiCL", "SCL", "L", "SiL", "SL", "Si", "LS", "S")

s1

numeric; horizontal coordinate (sand content 0–1) in the texture triangle system

s2

numeric; vertical coordinate (0–0.85) in the texture triangle system

Note

Texture by hand class can be converted to sand, silt, clay content fractions by using the TT2tri function. This function uses the v column in the USDA.TT.im (i.e. prior probability densities) to adjust for texture fraction combinations that are more probable.

Author(s)

Tomislav Hengl

References

Skaggs, T. H., Arya, L. M., Shouse, P. J., Mohanty, B. P., (2001) Estimating Particle-Size Distribution from Limited Soil Texture Data. Soil Science Society of America Journal 65 (4): 1038-1044.

See Also

FAO.SoilProfileCollection, soil.dom

Examples


## plot prior probabilities:
library(sp)
data(USDA.TT.im)
gridded(USDA.TT.im) <- ~s1+s2
spplot(USDA.TT.im["v"])

## Not run: library(soiltexture)
## convert textures by hand to sand, silt and clay:
TEXMHT <- c("CL","C","SiL","SiL","missing")
x <- TT2tri(TEXMHT)
x

## End(Not run)

[Package GSIF version 0.4-2 Index]