WPSworldgrids {GSIF}	R Documentation

Web Processing Service for worldgrids.org

Description

Overlay, subset and aggregation (wrapper) functions for the WPS worldgrids.org.

Usage


WPSworldgrids(layername, ...)

Arguments

layername
object

...
additional arguments

Details

WPS is implemented via the PyWPS module. The standard format for the images on the worldgrids.org repository is "GeoTiff". Complete grids (global coverage) can be obtained directly from worldgrids.org. We do not advise using the "bbox" argument to obtain grids that cover more than 30 percent of the global coverage.

Author(s)

Tomislav Hengl \& Hannes I. Reuter

References

text

text

See Also

landmask

Examples


## Not run: 
WPSworldgrids(layername="RNMGSP1", xcoord=10, ycoord=10)
WPSworldgrids(layername="TD1MOD1", bbox=matrix(c(-20,30,40,60), nrow=2))
# 1 km resolution:
WPSworldgrids(layername="LN1DSM3", bbox=matrix(c(7,51,9,54), nrow=2))

## End(Not run)

[Package GSIF version 0.1-1 Index]