afsp {GSIF}	R Documentation

Africa Soil Profiles Database

Description

A merge of the Africa Soil Profiles Database (AFSP) with 17,000+ geo-referenced legacy soil profile records, and AfSIS Sentinel Site database with 9000+ sampling locations.

Usage

data(afsp)

Format

The afsp data set contains two data frames — sites and horizons. Sites table contains the following columns:

SOURCEID

factor; unique label to help a user identify a particular site (ProfileID in the AFSP)

SOURCEDB

factor; source data base

LONWGS84

numeric; longitude in decimal degrees on the WGS84 datum (X_LonDD in the AFSP)

LATWGS84

numeric; latitude in decimal degrees on the WGS84 datum (Y_LatDD in the AFSP)

TIMESTRR

character; the date on which this particular soil was described or sampled (T_Year in the AFSP)

TAXGWRB

factor; abbreviated soil group based on the WRB classification system (WRB06rg in the AFSP)

TAXNUSDA

factor; Keys to Soil Taxonomy taxon name e.g. "Plinthic Udoxic Dystropept" (USDA in the AFSP)

BDRICM

numeric; depth to bedrock in cm

DRAINFAO

factor; drainage class based on the FAO guidelines for soil description: E (excessively drained), S (somewhat excessively drained), W (well drained), M (moderately well drained), I (somewhat poorly drained) and V (very poorly drained)

Horizons table contains the following columns:

SOURCEID

factor; a short label to help a user identify a particular site

UHDICM

numeric; upper horizon depth from the surface in cm

LHDICM

numeric; lower horizon depth from the surface in cm

MCOMNS

factor; Munsell color moist

ORCDRC

numeric; soil organic carbon content in permilles

PHIHOX

numeric; pH index measured in water solution

SNDPPT

numeric; weight percentage of the sand particles (0.05–2 mm)

SLTPPT

numeric; weight percentage of the silt particles (0.0002–0.05 mm)

CLYPPT

numeric; weight percentage of the clay particles (<0.0002 mm)

CRFVOL

numeric; volume percentage of coarse fragments (> 2 mm)

BLD

numeric;bulk density in tonnes per cubic-meter

CEC

numeric; Cation exchange capacity (fine earth fraction) in cmolc/kg

NTO

numeric; total N content in permille or g/kg

EMGX

numeric; exchangable Mg in cmolc/kg

Author(s)

The Africa Soil Profiles Database have been prepared by Johan Leenaars <johan.leenaars@wur.nl>. This is a subset of the original database that can be downloaded via www.isric.org. The AfSIS Sentinel Site database is one of the main deliverables of the Africa Soil Information Service project.

References

Leenaars, J.G.B. (2014) Africa Soil Profiles Database, Version 1.2. A compilation of geo-referenced and standardized legacy soil profile data for Sub Saharan Africa (with dataset). ISRIC report 2012/03. Africa Soil Information Service (AfSIS) project and ISRIC — World Soil Information, Wageningen, the Netherlands.

Africa Soil Information Service (http://africasoils.net)

Examples


## Not run: 
library(rgdal)
library(aqp)
library(sp)

data(afsp)
sites <- afsp$sites
coordinates(sites) <- ~ LONWGS84 + LATWGS84
proj4string(sites) <- "+proj=longlat +datum=WGS84"
## obtain country borders:
library(maps)
country.m = map('world', plot=FALSE, fill=TRUE)
IDs <- sapply(strsplit(country.m$names, ":"), function(x) x[1])
require(maptools)
country <- as(map2SpatialPolygons(country.m, IDs=IDs), "SpatialLines")
proj4string(country) = "+proj=longlat +datum=WGS84" 
## overlay and plot points and maps:
plot(country, col="darkgrey", xlim=c(-25.3,57.8), ylim=c(-34.8, 37.4))
points(sites, pch=21, bg="white", cex=.6, col="black")

## End(Not run)

[Package GSIF version 0.4-5 Index]