afss {GSIF}	R Documentation

Africa Sentinel Site data (wet chemistry)

Description

Africa soil sampled collected at AfSIS sentinel sites. Part of the African Land Degradation Surveillance Framework (LDSF) Field Database.

Usage

data(afss)

Format

The afss data set contains two data frames — sites and horizons. Sites table contains the following columns:

SOURCEID

factor; study soil profile ID (SPID in the LDSF Field Database)

CNTISOC

factor; ISO country code (CC in the LDSF Field Database)

LOCNAME

factor; sentinel site name referring to the closest populated place (Site in the LDSF Field Database)

LONWGS84

numeric; longitude in decimal degrees on the WGS84 datum (Lon in the LDSF Field Database)

LATWGS84

numeric; latitude in decimal degrees on the WGS84 datum (Lat in the LDSF Field Database)

Horizons table contains the following columns:

SOURCEID

factor; study soil profile ID (SPID in the LDSF Field Database)

SAMPLEID

factor; soil sample ID (SSID in the LDSF Field Database)

UHDICM

numeric; upper sampling depth from the surface in cm

LHDICM

numeric; lower sampling depth from the surface in cm

ORCCNS

numeric; soil organic carbon content in permille (SOC in the LDSF Field Database)

PHIHOX

numeric; pH index measured in water solution (pH in the LDSF Field Database)

EXB

numeric; Exchangeable bases in ppm (ExBas in the LDSF Field Database)

ALUM3S

numeric; Aluminium concentration in ppm (m3.Al in the LDSF Field Database)

ECAM3S

numeric; Exchangeable Ca in ppm (ExCa in the LDSF Field Database)

EXKM3S

numeric; Exchangeable K in ppm (ExK in the LDSF Field Database)

EMGM3S

numeric; Exchangeable Mg in ppm (ExMg in the LDSF Field Database)

ENAM3S

numeric; Exchangeable Mg in ppm (ExNa in the LDSF Field Database)

NITCNS

numeric; Exchangeable N as volumetric percentage (Total.Nitrogen in the LDSF Field Database)

SNDLDF

numeric; weight percentage of the sand particles (0.0002–0.05 mm; Sand in the LDSF Field Database)

afss.spec table contains the following columns:

SAMPLEID

factor; soil sample ID (SSID in the LDSF Field Database)

m4003.5

numeric; mid-infrared raw absorbance for denoted wavenumbers (Bruker-LTS instrument)

...

other raw absorbance values

Note

The soil samples from the sentinel site network were collected using three stage sampling scheme: a number of sentinel sites (about 10 by 10 kilometers in size) have been first selected over the whole continent (60 sites in 2012). Each sentinel site is stratied into 16 grid cells (each 1 square kilometer), and sampling cluster centroids are randomly located within the grid cells. Around each centroid, 10 sampling plots are randomly located and soil samples taken (or 160 points per sentinel site). At each sampling plot, soil samples were taken at two depths: 0–20, and 20–50 cm. The block size of each sampling plot is 1000 square meters or 0.1 ha.

Author(s)

Africa Soil Information Service (http://africasoils.net)

References

Vagen, T., Shepherd, K. D., Walsh, M. G., Winowiecki, L., Desta, L. T., & Tondoh, J. E. (2010) AfSIS technical specifications: soil health surveillance. World Agroforestry Centre, Nairobi, Kenya.

Examples


## Not run: 
library(sp)
library(rgdal)

data(afss)
xy <- afss$sites
coordinates(xy) <- ~ LONWGS84 + LATWGS84
proj4string(xy) <- "+proj=longlat +datum=WGS84"
xy$PHIHOX_d1 <- merge(xy@data, afss$horizons[afss$horizons$UHDICM==0,], all.x=TRUE)$PHIHOX

## plot points in Google Earth:
library(plotKML)
data(R_pal)
plotKML(xy["PHIHOX_d1"], colour_scale=R_pal[["pH_pal"]])

## obtain country borders:
con.admin <- url("http://gsif.isric.org/lib/exe/fetch.php?media=admin.af.rda")
load(con.admin)
proj4string(admin.af) <- get("ref_CRS", envir = plotKML.opts)
admin.af <- as(admin.af, "SpatialLines") 
## overlay and plot points and maps:
plot(admin.af, col="darkgrey", xlim=xy@bbox[1,], ylim=xy@bbox[2,])
points(xy, pch=21, bg="white", cex=.6, col="black")

## obtain MIR measurements:
con <- url("http://gsif.isric.org/lib/exe/fetch.php?media=afss.spec.rda")
load(con)
str(afss.spec)
## See "soil.spec" package for further processing...

## End(Not run)

[Package GSIF version 0.4-3 Index]