as.geosamples {GSIF}	R Documentation

Converts an object of class "SpatialProfileCollection" to a standardized data frame

Description

Converts an object of class "SpatialProfileCollection" to an object of class "geosamples" with all measurements broken separately as individual records. Geo-records are standardized universal spatially and temporally referenced samples from the Earth's surface.

Usage


as.geosamples(obj)

Arguments

obj
object of class "SoilProfileCollection"

Details

Returns an object of type "geosamples" with the following slots:

registry

character; an unique URI defining the global registry with all metadata and attribute lists (domains)

methods

data.frame; a table with method names ("methodid"), a detailed description of each method ("description"), measurement units or levels ("units"), and associated detection limits ("detectionLimit")

data

data.frame; a standardized table with fixed column names: "sampleid" (universal ID; as specified in the data registry service), "producerid" (producer's ID; usually site ID and horizon ID or sequence number), "longitude", "latitude", "locationError", "TimeSpan.begin", "TimeSpan.end", "altitude", "altitudeError", "altitudeMode", "dimension", "volume", "value" (measured value), "methodid", "measurementError"

The column names in the data slot largely reflect the KML schema elements. For more information about what each column means, refer to the KML reference.
Many columns required by the "geosamples" class can be missing and will result in NA values. These will likely need to be inserted afterwards, within or outside R.

Note

The basic idea of this class is to outsource the metadata management to some global publicly available service. Georecord class can be used to prepare geological, hydrological, geochemical, biodiversity, soil science and similar samples. Geological and soil samples can be registered via the geosamples.org, for example, so that the "sampleid" corresponds to the IGSN (unique sample identifier). The methods used in the field of soil science could be all implemented via the SoilLM scheme. "producerid" column is recommended as it allows linking to the original data set.

Author(s)

Tomislav Hengl and Hannes I. Reuter

See Also

as.data.frame, geosamples-class

[Package GSIF version 0.1-1 Index]