autopredict-methods {GSIF}	R Documentation

Auto predict numeric or factor type variables

Description

Fits either geostatistical model via the fit.gstatModel function (in the case of numeric variable) or a multinomial logistic regression model via the spmultinom function (factor-type variable) and generates predictions.

Usage


## S4 method for signature 'SpatialPointsDataFrame,SpatialPixelsDataFrame'
autopredict(target,covariates, 
     auto.plot=TRUE, ...)

Arguments

target
object of class "SpatialPointsDataFrame" containing observations of the target variable

covariates
object of class "SpatialPixelsDataFrame"; spatial covariates

auto.plot
logical; specifies whether to immediately plot the data via the plotKML function

...
other optional arguments that can be passed to fit.gstatModel or spmultinom

Author(s)

Tomislav Hengl

See Also

fit.gstatModel

Examples


## Ebergotzen data:
library(sp)
library(gstat)
library(ranger)
library(plotKML)

## load input data:
data(eberg)
eberg <- eberg[runif(nrow(eberg))<.1,]
coordinates(eberg) <- ~X+Y
proj4string(eberg) <- CRS("+init=epsg:31467")
data(eberg_grid)
gridded(eberg_grid) <- ~x+y
proj4string(eberg_grid) <- CRS("+init=epsg:31467")

## predict sand content:
SNDMHT_A <- autopredict(eberg["SNDMHT_A"], eberg_grid,
   auto.plot=FALSE)
plot(SNDMHT_A)

## predict soil types:
soiltype <- autopredict(eberg["soiltype"], eberg_grid, 
  auto.plot=FALSE)
spplot(soiltype@predicted)

[Package GSIF version 0.5-1 Index]