edgeroi {GSIF}	R Documentation

The Edgeroi Data Set

Description

Soil samples and covariate layers for the Edgeroi area in NSW, Australia (ca 1500 square-km).

Usage

data(edgeroi)

Format

The edgeroi data set contains two data frames — sites and horizons. Sites table contains the following columns:

SOURCEID

factor; unique label to help a user identify a particular site (ID in the NatSoil)

LONGDA94

numeric; longitude in decimal degrees on the GDA94 datum

LATGDA94

numeric; latitude in decimal degrees on the GDA94 datum

TAXGAUC

factor; Australian Great Soil Groups (GSG; see details)

NOTEOBS

character; free-form observation notes

Horizons table contains the following columns:

SOURCEID

factor; unique identifier used in the NatSoil DB

LSQINT

integer; a layer sequence number 1 to N

HZDUSD

factor; horizon designation (primary letter)

UHDICM

numeric; lower horizon depth from the surface in cm

LHDICM

numeric; upper horizon depth from the surface in cm

CLYPPT

numeric; weight percentage of the clay particles (<0.0002 mm)

SNDPPT

numeric; weight percentage of the silt particles (0.0002–0.05 mm)

SLTPPT

numeric; weight percentage of the sand particles (0.05–2 mm)

PHIHO5

numeric; pH index measured in water solution(ph_h2o in the NSCD)

ORCDRC

numeric; soil organic carbon content in permille

The edgeroi.grids data frame contains a list of covariates at 250 m resolution:

DEMSRT5

numeric; SRTM DEM

TWISRT5

numeric; SAGA Topographic Wetness Index based on the SRTM DEM

PMTGEO5

factor; parent material class based on the National Geological map at scale 1:250,000 — sand with minor silty sand ("Qd"), alluvium gravel, sand, silt, clay ("Qrs"), quartz sandstone obscured by quartenary sands ("Qrt/Jp"), quartz sandstone obscured by talus material ("Qrt/Rn"), basalt obscured by talus material ("Qrt/Tv"), mottled clay, silt, sandstone and gravel ("Ts"), and basalt, dolerite, trachyte, techenite ("Tv")

EV1MOD5

numeric; first principal component of the MODIS EVI (MOD13Q1) time series data (year 2011)

EV2MOD5

numeric; second principal component of the MODIS EVI (MOD13Q1) time series data (year 2011)

EV3MOD5

numeric; third principal component of the MODIS EVI (MOD13Q1) time series data (year 2011)

x

numeric; x-coordinate in the GDA94 / MGA zone 55

y

numeric; y-coordinate in the GDA94 / MGA zone 55

The edgeroi.grids100 data frame contains a list of covariates at 100 m resolution prepared for the study area:

LNUABS6

factor; Australian National scale land use data

MVBSRT6

numeric; SAGA GIS Multi-resolution Index of Valley Bottom Flatness based on the SRTM DEM

TI1LAN6

numeric; principal component 1 for the Landsat band 7 (thermal) based on three periods of the Global Land Survey Landsat images (GLS1990, GLS2000, GLS2005)

TI2LAN6

numeric; principal component 2 for the Landsat band 7 (thermal) based on three periods of the Global Land Survey Landsat images (GLS1990, GLS2000, GLS2005)

PCKGAD6

numeric; percentage of Potassium estimated based on the gamma radiometrics radmap09 (GADDS)

RUTGAD6

numeric; ratio Uranium over Thorium estimated based on the gamma radiometrics radmap09 (GADDS)

PCTGAD6

numeric; parts per million of Thorium estimated based on the gamma radiometrics radmap09 (GADDS)

x

numeric; x-coordinate in the GDA94 / MGA zone 55

y

numeric; y-coordinate in the GDA94 / MGA zone 55

Details

The Edgeroi is one of the standard soil data sets used to test soil mapping methods in Australia. Out of 359 profiles, 210 sites were sampled on a systematic, equilateral triangular grid with a spacing of 2.8 km between sites, the other sites are distributed more irregularly or on transects. The data set is described in detail in Malone et al. (2010) and McGarry et al. (1989). The edgeroi contains only a subset of the original NatSoil records. Observed soil classes for TAXGAUC are (alphabetically): Alluvial soil ("A"), Brown clay ("BC"), Black earth ("BE"), Earthy sand ("ES"), Grey clay ("GC"), Grey earth ("GE"), No suitable group ("NSG"), Prairie soil ("PS"), Rendzina ("R"), Red-brown earth ("RBE"), Red clay ("RC"), Red earth ("RE"), Red podzolic soil ("RP"), Solodic soil ("SC"), Soloth ("SH"), Solonchak ("SK"), Siliceous sand ("SS"), and Solonetz ("SZ").

Note

The Landsat images and SRTM DEM have been obtained from the Global Land Cover Facility. Scanned geology map (paper sheets) has been obtained from the Geoscience Australia, then georeferenced and rasterized to 250 m resolution. The land use map has been obtained from the Australian Collaborative Land Use and Management program. The Radiometric Map of Australia grids has been downloaded using the Geophysical Archive Data Delivery System (GADDS) on the Australian Government's Geoscience Portal (Mitny et al, 2009).
Listed gridded layers follow a standard naming convention used by WorlGrids.org (the standard 8.3 filename convention with at most eight characters): first three letter are used for the variable type e.g. DEM (digital elevation model); the next three letters represent the data source or collection method e.g. SRT (SRTM mission); the 6th character is the effective scale e.g. 5 indicates the 5th standard scale i.e. 1/600 decimal degrees (in this case 250 m).

Author(s)

The original detailed profile description and laboratory analysis was funded by a Cotton Research and Development Corporation project in the mid-late 1980's by the CSIRO Division of Soils and available via the NatSoil DB. The gamma radiometrics images are property of the NSW Department of Primary Industries — Mineral Resources.

References

Malone, B.P., McBratney, A.B., Minasny, B. (2010) Mapping continuous depth functions of soil carbon storage and available water capacity. Geoderma 154, 138-152.

McGarry, D., Ward, W.T., McBratney, A.B. (1989) Soil Studies in the Lower Namoi Valley: Methods and Data. The Edgeroi Data Set. (2 vols) (CSIRO Division of Soils: Adelaide).

Minty, B., Franklin, R., Milligan, P., Richardson, L.M., and Wilford, J., (2009) The Radiometric Map of Australia. Exploration Geophysics, 40(4), 325-333.

Examples


library(rgdal)
library(aqp)
library(sp)

data(edgeroi)
edgeroi$sites[edgeroi$sites$SOURCEID=="399_EDGEROI_ed095_1",]
edgeroi$horizons[edgeroi$horizons$SOURCEID=="399_EDGEROI_ed095_1",]
## spPoints:
sites <- edgeroi$sites
coordinates(sites) <- ~ LONGDA94 + LATGDA94
proj4string(sites) <- CRS("+proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +no_defs")
## Not run: ## plot points and grids:
pnts <- list("sp.points", spTransform(sites, CRS("+init=epsg:28355")), pch="+", col="black")
## load the 250 m grids:
download.file("http://worldgrids.org/rda/edgeroi.grids.rda", "edgeroi.grids.rda")
load("edgeroi.grids.rda")
gridded(edgeroi.grids) <- ~x+y
proj4string(edgeroi.grids) <- CRS("+init=epsg:28355")
spplot(edgeroi.grids[1], sp.layout=pnts)
## load the 100 m grids:
download.file("http://worldgrids.org/rda/edgeroi.grids100.rda", "edgeroi.grids100.rda")
load("edgeroi.grids100.rda")
gridded(edgeroi.grids100) <- ~x+y
proj4string(edgeroi.grids100) <- CRS("+init=epsg:28355")
spplot(edgeroi.grids100["TI1LAN6"], sp.layout=pnts)

## End(Not run)

[Package GSIF version 0.3-3 Index]