fit.gstatModel-methods {GSIF}	R Documentation

Methods to fit a regression-kriging model

Description

Tries to automatically fit a 2D or 3D regression-kriging model for a given set of points (object of type "SpatialPointsDataFrame" or "geosamples") and covariates (object of type "SpatialPixelsDataFrame"). It first fits a regression model (Generalized Linear Model) following the formulaString, then fits variogram for residuals usign the fit.variogram method from the gstat package. Creates an output object of class gstatModel-class.

Usage

 
## S4 method for signature 'SpatialPointsDataFrame,formula,SpatialPixelsDataFrame'
fit.gstatModel(observations, formulaString, covariates, 
     method = list("GLM", "cubist", "HB")[[1]], 
     dimensions = list("3D", "2D", "2D+T", "3D+T")[[1]],
     family = gaussian, stepwise = TRUE, vgmFun = "Exp", rvgm = NULL,
     subsample = 5000, ...)
## S4 method for signature 'geosamples,formula,SpatialPixelsDataFrame'
fit.gstatModel(observations, formulaString, covariates, 
     methodid, method = list("GLM", "cubist", "HB")[[1]], 
     dimensions = list("3D", "2D", "2D+T", "3D+T")[[1]],
     family = gaussian, stepwise = TRUE, 
     vgmFun = "Exp", rvgm = NULL, subsample = 5000, ...)
## S4 method for signature 'geosamples,formula,list'
fit.gstatModel(observations, formulaString, covariates, 
     methodid, method = list("GLM", "cubist", "HB")[[1]], 
     dimensions = list("3D", "2D", "2D+T", "3D+T")[[1]],
     family = gaussian, stepwise = TRUE, 
     vgmFun = "Exp", rvgm = NULL, subsample = 5000, ...)
## S4 method for signature 'geosamples,list,list'
fit.gstatModel(observations, formulaString, covariates, 
     methodid, method = list("GLM", "cubist", "HB")[[1]], 
     dimensions = list("3D", "2D", "2D+T", "3D+T")[[1]],
     family = gaussian, stepwise = TRUE, 
     vgmFun = "Exp", rvgm = NULL, subsample = 5000, ...)

Arguments

observations
object of type "SpatialPointsDataFrame" or "geosamples-class"

formulaString
object of type "formula" or a list of formulas

covariates
object of type "SpatialPixelsDataFrame", or list of grids

methodid
character; method name in the case of using the "geosamples-class"

method
character; family of methods considered e.g. "GLM"

dimensions
character; "3D", "2D", "2D+T", "3D+T" models

family
character string defyning the GLM family (for more info see stats::glm)

stepwise
specifies whether to run step-wise regression on top of GLM to get an optimal subset of predictors

vgmFun
variogram function ("Exp" by default)

rvgm
residual variogram model can be passed from gstat, in which case fit.gstatModel will not try to fit the variogram

subsample
integer; maximum number of observations to be taken for model fitting (to speed up variogram fitting)

...
other optional arguments that can be passed to glm and/or fit.variogram

Details

The GLM method by default assumes that the target variable follows a normal distribution family = gaussian. Other possible families are:

normal distribution

family = gaussian (default setting)

log-normal distribution

family = gaussian(log)

binomial variable

family = binomial(logit)

variable following a poisson distribution

family = poisson(log)

Note

Residuals will be checked for normality and problems reported by default. The warning messages should be taken with care, as when the sample size is small, even big departures from normality will not be reported; when the sample size is large, even the smallest deviation from normality might lead to a warning. Likewise, if the variogram fitting fails, consider fitting a variogram manually or using the fit.vgmModel method.

Author(s)

Tomislav Hengl and Gerard B.M. Heuvelink

References

chapter 8 “Interpolation and Geostatistics” in Bivand, R., Pebesma, E., Rubio, V., (2008) Applied Spatial Data Analysis with R. Use R Series, Springer, Heidelberg, pp. 378.

Hengl, T. (2009) A Practical Guide to Geostatistical Mapping, 2nd Edt. University of Amsterdam, www.lulu.com, 291 p.

See Also

gstatModel-class, fit.regModel, geosamples-class, stats::glm, gstat::fit.variogram

Examples


# 2D model:
library(sp)
data(meuse)
coordinates(meuse) <- ~x+y
proj4string(meuse) <- CRS("+init=epsg:28992")
# load grids:
data(meuse.grid)
coordinates(meuse.grid) <- ~x+y
gridded(meuse.grid) <- TRUE
proj4string(meuse.grid) <- CRS("+init=epsg:28992")
# fit a model with a gaussian log-link:
omm <- fit.gstatModel(meuse, om~dist, meuse.grid, family = gaussian(log))
# it was succesful
omm@regModel
om.sp <- SpatialPointsDataFrame(omm@sp, data = omm@regModel$model)
## Not run: # plot a variogram:
plot(variogram(om~1, om.sp))

## End(Not run)

# 3D model:
library(aqp)
library(plyr)
library(splines)
data(eberg)
# list columns of interest:
s.lst <- c("ID", "soiltype", "TAXGRSC", "X", "Y")
h.lst <- c("UHDICM","LHDICM","SNDMHT","SLTMHT","CLYMHT")
sel <- runif(nrow(eberg))<.05
# get sites table:
sites <- eberg[sel,s.lst]
# get horizons table:
horizons <- getHorizons(eberg[sel,], idcol="ID", sel=h.lst)
# create object of type "SoilProfileCollection"
eberg.spc <- join(horizons, sites, type='inner')
depths(eberg.spc) <- ID ~ UHDICM + LHDICM
site(eberg.spc) <- as.formula(paste("~", paste(s.lst[-1], collapse="+"), sep=""))
coordinates(eberg.spc) <- ~X+Y
proj4string(eberg.spc) <- CRS("+init=epsg:31467")
# convert to geosamples:
eberg.geo <- as.geosamples(eberg.spc)
# covariates:
data(eberg_grid)
gridded(eberg_grid) <- ~x+y
proj4string(eberg_grid) <- CRS("+init=epsg:31467")
glm.formulaString = as.formula(paste("observedValue ~ ", 
  paste(names(eberg_grid), collapse="+"), "+ ns(altitude, df=4)"))
SNDMHT.m <- fit.gstatModel(observations=eberg.geo, 
  glm.formulaString, covariates=eberg_grid, methodid="SNDMHT")
# problems with the variogram?
SNDMHT.m@vgmModel

[Package GSIF version 0.2-4 Index]