fit.regModel-methods {GSIF}	R Documentation

Fits a regression model to spatial data

Description

Fits a regression model (e.g. GLM) and variogram for the response residuals using the default settings.

Usage


## S4 method for signature 'formula,data.frame,SpatialPixelsDataFrame,character'
fit.regModel(formulaString,
      rmatrix, predictionDomain, method = list("GLM", "CART", "HB")[[1]], 
      dimensions = NULL, family=gaussian, stepwise=TRUE, rvgm=NULL, 
      vgmFun="Exp", subsample = 5000, ...)

Arguments

formulaString
object of class "formula" — regression model

rmatrix
object of class "data.frame"; regression matrix produced as a result of spatial overlay

predictionDomain
object of class "SpatialPixelsDataFrame"; spatial domain of interest

method
character; family of methods considered e.g. "GLM"

dimensions
character; "3D", "2D", "2D+T", "3D+T" models

family
family to be passed to the glm (see examples below)

stepwise
specifies whether to run step-wise regression on top of GLM to get an optimal subset of predictors

rvgm
residual variogram (assumed to be unknown)

vgmFun
character; variogram function ("Exp" by default)

subsample
integer; the threshold size at which the method will randomly subset the observations to speed up the model fitting

...
other optional arguments that can be passed to glm and/or gstat::fit.variogram

Details

Produces an object of class "gstatModel" that contains: (1) fitted GLM, (2) fitted variogram, and (c) object of class "SpatialPoints" with observation locations. To combine overlay and model fitting operations, use fit.gstatModel.

Author(s)

Tomislav Hengl

See Also

fit.gstatModel, stats::glm, gstat::fit.variogram

Examples


# universal kriging with Meuse data:
data(meuse)
coordinates(meuse) <- ~x+y
proj4string(meuse) <- CRS("+init=epsg:28992")
data(meuse.grid)
gridded(meuse.grid) <- ~x+y
proj4string(meuse.grid) <- CRS("+init=epsg:28992")
# produce a regression matrix:
index <- overlay(meuse.grid, meuse)
sel <- !is.na(index)
ov <- cbind(data.frame(meuse[sel,"om"]), meuse.grid[index[sel],])
# fit a model:
m <- fit.regModel(om~dist+ffreq, rmatrix=ov, meuse.grid, method="GLM", family=gaussian(log))
m@regModel
m@vgmModel

[Package GSIF version 0.2-5 Index]