fit.vgmModel-methods {GSIF}	R Documentation

Fits a 2D or 3D variogram model to spatial data

Description

Fits a 2D or 3D variogram model based on a regression matrix and spatial domain of interest.

Usage


## S4 method for signature 'formula,data.frame,SpatialPixelsDataFrame'
fit.vgmModel(formulaString,
      rmatrix, predictionDomain, vgmFun = "Exp", 
      dimensions = list("3D", "2D", "2D+T", "3D+T")[[1]],
      anis = NULL, subsample = nrow(rmatrix), ...)

Arguments

formulaString
object of class "formula" — regression model

rmatrix
object of class "data.frame"; regression matrix produced as a result of spatial overlay

predictionDomain
object of class "SpatialPixelsDataFrame"; spatial domain of interest

vgmFun
character; variogram function ("Exp" by default)

dimensions
character; "3D", "2D", "2D+T", "3D+T" models

anis
vector containing 2, 5 or more anisotropy parameters; see gstat::vgm for more info

subsample
integer; size of the subset

...
other optional arguments that can be passed to gstat::fit.variogram

Details

It will try to fit a variogram to multidimensional data. If the data set is large, this process can be time-consuming, hence one way to speed up fitting is to subset the regression matrix using the subsample argument (random subsetting).

Author(s)

Tomislav Hengl

See Also

fit.regModel, fit.gstatModel, gstat::fit.variogram

Examples


# fit variogram to the Meuse data:
data(meuse)
coordinates(meuse) <- ~x+y
proj4string(meuse) <- CRS("+init=epsg:28992")
data(meuse.grid)
gridded(meuse.grid) <- ~x+y
proj4string(meuse.grid) <- CRS("+init=epsg:28992")
# produce a regression matrix:
index <- overlay(meuse.grid, meuse)
sel <- !is.na(index)
ov <- data.frame(meuse.grid[index[sel],])
ov$om <- meuse@data[sel,"om"]
# fit a model:
v <- fit.vgmModel(om~1, rmatrix=ov, meuse.grid, dimensions="2D")
plot(variogram(om ~ 1, meuse[!is.na(meuse$om),]), v$vgm)

[Package GSIF version 0.2-5 Index]