geochm {GSIF}	R Documentation

NGS database samples for Indiana State

Description

A subset of the National Geochemical Survey (NGS) samples covering the Indiana and Illinois State. Contains a total of 2681 point samples.

Usage

data(geochm)

Format

Data frame; contains the following columns:

REC_NO

factor; unique record identifier

DATASET

factor; abbreviated dataset group e.g. "AK+MI"

TYPEDESC

factor; abbreviated description of sample type: stream, pond, spring, soil etc

COLL_DATE

integer; sampling date

LONGITUDE

numeric; longitude in decimal degrees (NAD27 datum)

LATITIUDE

numeric; latitude in decimal degrees (NAD27 datum)

DATUM

factor; geodetic datum if different from NAD83

RELIEF

factor; relief in drainage basin from which sample was collected

FORMATION

factor; code or name of geologic formation in which sample area was located

ROCK_TYPE

factor; rock type in area of sample collection e.g. "carbonate"

SOIL_HORIZ

factor; soil horizon from which the sample was collected

COLOR

factor; observed color of powdered sample during splitting

MEDIUM

factor; sample medium — rock, sediment, standard, or unknown

SOURCE

factor; geological source of the sample medium that was collected e.g. "Beach"

AS_ICP40

numeric; As (ppm) by Inductively Coupled Plasma Spectrometry (ICP) after acid dissolution

CD_ICP40

numeric; Cd (ppm)

CR_ICP40

numeric; Cr (ppm)

CU_ICP40

numeric; Cu (ppm)

NI_ICP40

numeric; Ni (ppm)

ZN_ICP40

numeric; Zn (ppm)

AS_AA

numeric; As (ppm) by Hydride Atomic Absorption

HG_AA

numeric; Hg (ppm) by Hydride Atomic Absorption

PB_ICP40

numeric; Pb (ppm)

C_TOT

numeric; total carbon (weight percentage) by combustion

C_ORG

numeric; organic carbon (weight percentage) as a difference between C_TOT and C_CO3

C_CO3

numeric; carbonate carbon (weight percentage) by Coulometric Titration

S_TOT

numeric; total sulfur (weight percentage) by combustion

Note

Negative values of the heavy metal concetrations indicate a determination that is below the limit of detection for the analytic method used. The magnitude of the negative number indicates the detection limit. For example, -10 ppm means the result should be regarded as < 10 ppm.

Author(s)

National Geochemical Survey database is maintaned by the USGS National Geochemical Survey Team (contact: Peter Schweitzer). This subset has been prepared for the purpose of testing various geostatistical mapping algoriths by Tomislav Hengl (tom.hengl@wur.nl).

References

The National Geochemical Survey Team, (2008) The National Geochemical Survey: database and documentation. U.S. Geological Survey Open-File Report 2004-1001, U.S. Geological Survey, Reston VA.

National Geochemical Survey database (http://tin.er.usgs.gov/geochem/)

Examples


library(sp)

# Load the NGS data:
data(geochm)
coordinates(geochm) <- ~LONGITUDE+LATITUDE
proj4string(geochm) <- CRS("+proj=longlat +ellps=clrk66 +datum=NAD27 +no_defs")
## Not run: 
require(plotKML)
data(SAGA_pal)
# replace the missing values with half the detection limit:
geochm$PB_ICP40 <- ifelse(geochm$PB_ICP40 < 0, 2, geochm$PB_ICP40) 
shape = "http://maps.google.com/mapfiles/kml/pal2/icon18.png"
kml(geochm, shape = shape, colour = log1p(PB_ICP40), labels = "", 
    colour_scale = SAGA_pal[[1]], kmz = TRUE)

## End(Not run)

[Package GSIF version 0.4-0 Index]