geosamples-class {GSIF}	R Documentation

A class for spatially and temporally referenced samples

Description

A class for spatially and temporally referenced samples with fixed column names (standardized geosamples). Corresponds to the point "Placemark" in the KML schema.

Slots

registry:

object of class "character"; URI of the online registry i.e. the URL where the "producerid" column can be linked to all other connected metadata

methods:

object of class "data.frame"; a table with method names ("methodid"), a one sentence description of each method ("description"), measurement units or levels ("units"), and associated detection limits ("detectionLimit")

data:

object of class "data.frame"; a standardized table with fixed column names: "observationid" (unique observation ID; as specified in the data registry service), "sampleid" (producer's ID; usually site ID and horizon ID or sequence number), "longitude" (longitude on the WGS84 ellipsoid), "latitude" (latitude on the WGS84 ellipsoid), "locationError" (error radius in meters), "TimeSpan.begin" (begin of the measurement period), "TimeSpan.end" (end of the measurement period), "altitude" (height above ground or above the sea level in meters), "altitudeMode" (one of the KML schema altitude modes), "sampleArea" (spatial support in square meters), "sampleThickness" (thickness of horizons in meters or vertical support), "observedValue" (measured value), "methodid" (method name; see methods table), "measurementError" (estimated measurement error for that specific observation)

The column names in the data slot largely reflect the KML schema elements. Geosamples are interoperable with the OGC Observations and measurements specifications, but do not necessarily contain all required fields (i.e. there is no validity check for the OGC specifications). Geosamples-class can be used to store and manipulate geological, hydrological, geochemical, biodiversity, soil science and similar field samples near or below land surface. Geological and soil samples can also be registered via the geosamples.org, in which case the "observationid" will correspond to the unique sample identifier. "sampleid" column allows linking geosamples to the original ID's.

Methods

show

signature(obj = "geosamples"): summarize object by listing methods, total number of observations, total area covered etc.

subset

signature(obj = "geosamples"): subset to a single variable type; returns a data frame

over

signature(x = "SpatialPixelsDataFrame" or "RasterStack", y = "geosamples"): overlay geosamples and spatial pixels

write.data

signature(obj = "geosamples"): write geosamples to an external format e.g. GeoEAS

Author(s)

Tomislav Hengl and Hannes I. Reuter

References

Dyson, E., (2003) Online Registries: The DNS and Beyond.... Edventure, Vol 21(8).

International Geo Sample Number (http://en.wikipedia.org/wiki/International_Geo_Sample_Number)

KML Reference (https://developers.google.com/kml/documentation/kmlreference)

OGC Observations and Measurements standard (http://www.opengeospatial.org/standards/om/)

SESAR, the System for Earth Sample Registration (http://www.geosamples.org)

See Also

as.geosamples

[Package GSIF version 0.4-3 Index]