geosamples-class {GSIF}	R Documentation

A class for spatially and temporally referenced samples

Description

A class for spatially and temporally referenced samples with fixed column names (standardized geosamples). Corresponds to the point "Placemark" in the KML schema.

Slots

registry:

object of class "character"; URI of the online registry i.e. the URL where the "producerid" column can be linked to all other connected metadata

methods:

object of class "data.frame"; a table with method names ("methodid"), a one sentence description of each method ("description"), measurement units or levels ("units"), and associated detection limits ("detectionLimit")

data:

object of class "data.frame"; a standardized table with fixed column names: "observationid" (universal ID; as specified in the data registry service), "sampleid" (producer's ID; usually site ID and horizon ID or sequence number), "longitude", "latitude", "locationError" (error radius in meters), "TimeSpan.begin", "TimeSpan.end", "altitude", "altitudeMode", "volume", "value" (measured value), "methodid", "measurementError"

The column names in the data slot largely reflect the KML schema elements. For more information refer to the KML reference. Geosamples-class can be used to store and manipulate geological, hydrological, geochemical, biodiversity, soil science and similar field samples. Geological and soil samples can be registered via the geosamples.org, in which case the "observationid" will correspond to the unique sample identifier (IGSN). "producerid" column allows linking geosamples to the original ID's.

Methods

show

signature(obj = "geosamples"): summarize object by listing methods, total number of observations, total area covered etc.

subset

signature(obj = "geosamples"): subset to a single variable type; returns a data frame

overlay

signature(covariates = "SpatialPixelsDataFrame", observations = "geosamples"): overlay geosamples and spatial pixels

write.data

signature(obj = "geosamples"): write geosamples to an external format e.g. GeoEAS

Author(s)

Tomislav Hengl and Hannes I. Reuter

References

Dyson, E., (2003) Online Registries: The DNS and Beyond.... Edventure, Vol 21(8).

International Geo Sample Number (http://en.wikipedia.org/wiki/International_Geo_Sample_Number)

KML Reference (https://developers.google.com/kml/documentation/kmlreference)

SESAR, the System for Earth Sample Registration (http://www.geosamples.org)

See Also

as.geosamples

[Package GSIF version 0.2-1 Index]