getSpatialTiles {GSIF}	R Documentation

Get a list of tiles (regular blocks)

Description

Creates a list of tiles ("SpatialPolygons") for a given spatial domain i.e. extent.

Usage


## S4 method for signature 'Spatial'
getSpatialTiles(obj, block.x, block.y = block.x, 
    overlap.percent = 0, limit.bbox = TRUE, return.SpatialPolygons = TRUE)

Arguments

obj
object of class "Spatial*"

block.x
numeric; size of block in x-direction (meters or corresponding mapping units)

block.y
numeric; size of block in y-direction (meters or corresponding mapping units)

overlap.percent
numeric; percentage overlap (must be a positive number)

limit.bbox
logical; specifies whether to limit the extent of tiles to the bounding box only

return.SpatialPolygons
logical; specifies whether to return a list of tiles as "SpatialPolygons" or a data frame with bounding box coordinates

Details

The first output tile starts by default at the lower left corner. getSpatialTiles-method can only be used to generate regular tiles.

Value

Returns a list of tiles either as a list of "SpatialPolygons" or a data frame with with bounding box coordinates.

Author(s)

Tomislav Hengl

See Also

tile, sp::spsample

[Package GSIF version 0.2-5 Index]