glm.sp {GSIF}	R Documentation

Fits a GLM to spatial data

Description

Fits a GLM and variogram for the residuals using the default settings used in the gstat and glm packages.

Usage


glm.sp(formulaString, rmatrix, predictionDomain, family=gaussian, stepwise=TRUE, rvgm=NULL, vgmFun="Exp", type, ...)

Arguments

formulaString
object of class "formula" — regression model

rmatrix
object of class "data.frame"; regression matrix produced as a result of spatial overlay

predictionDomain
object of class "SpatialPixelsDataFrame"; spatial domain of interest

family
family to be passed to the glm (see examples below)

stepwise
specifies whether to run step-wise regression on top of GLM to get an optimal subset of predictors

rvgm
residual variogram (assumed to be unknown)

vgmFun
character; variogram function ("Exp" by default)

type
character; spatial object type (e.g. "SpatialPointsDataFrame" or "geosamples")

...
other optional arguments that can be passed to glm and/or fit.variogram

Details

Produces a list of three objects: (1) fitted GLM, (2) fitted variogram, and (c) object of class "SpatialPoints" with observation locations. To combine overlay and model fitting operations, use fit.gstatModel.

Author(s)

Tomislav Hengl

See Also

fit.gstatModel, stats::glm, gstat::fit.variogram

Examples


# universal kriging with Meuse data:
data(meuse)
coordinates(meuse) <- ~x+y
proj4string(meuse) <- CRS("+init=epsg:28992")
data(meuse.grid)
gridded(meuse.grid) <- ~x+y
proj4string(meuse.grid) <- CRS("+init=epsg:28992")
# produce a regression matrix:
index <- overlay(meuse.grid, meuse)
sel <- !is.na(index)
ov <- cbind(data.frame(meuse[sel,"om"]), meuse.grid[index[sel],])
# fit a model:
m <- glm.sp(om~dist, rmatrix=ov, meuse.grid, family=gaussian(log), type="SpatialPointsDataFrame")
m$regModel
m$vgmModel

[Package GSIF version 0.2-2 Index]