gstatModel-class {GSIF}	R Documentation

A class for a linear geostatistical model

Description

A class containing fitted parameters of a linear geostatistical model. It contains regression (GLM) and the kriging part (variogram).

Slots

regModel:

object of class "glm"; output of fitting a generalized linear model (GLM)

vgmModel:

object of class "data.frame"; the fitted gstat variogram model parameters containing

sp:

object of class "SpatialPointsDataFrame"; observation locations

Methods

summary

signature(obj = "gstatModel"): summarize object by showing the mapping accuracy (cross-validation) and the amount of variation explained by the model

predict

signature(obj = "gstatModel"): makes predictions for a set of given predictionLocations (gridded maps)

Note

"SpatialPredictions" saves results of predictions for a single target variable, which can be of type numeric or factor. Multiple variables can be combined into a list. When using nsim argument with the predict method, the output result will be of type:

plotKML::RasterBrickSimulations-class

i.e. N number of equiprobable realizations. To generate an object of type:

SpatialPredictions-class

use nsim = 0.

Author(s)

Tomislav Hengl

See Also

plotKML::SpatialPredictions-class, plotKML::RasterBrickSimulations-class, gstat::gstat, stats::glm

Examples

 
# load observations:
data(meuse)
coordinates(meuse) <- ~x+y
proj4string(meuse) <- CRS("+init=epsg:28992")
# load grids:
data(meuse.grid)
coordinates(meuse.grid) <- ~x+y
gridded(meuse.grid) <- TRUE
proj4string(meuse.grid) <- CRS("+init=epsg:28992")
## Not run: # fit a model:
omm <- fit.gstatModel(meuse, formulaString = om~dist, covariates = meuse.grid)
show(omm@regModel)
# produce SpatialPredictions:
om.rk <- predict(omm, predictionLocations = meuse.grid)
# overview of the prediction process:
show(om.rk)
# plot the whole project and open in Google Earth:
library(plotKML)
plotKML(om.rk)

## End(Not run)

[Package GSIF version 0.2-0 Index]