hoosier {GSIF}	R Documentation

Hoosier National Forest soil covariates at 1 km resolution

Description

A list of standard soil mapping covariates prepared for the souther Indiana state (Hoosier National Forest area; 33,581 grid nodes). Projected in the Albers equal-area projection system that is commonly used to visualize the contiguous 48-state area.

Usage

data(hoosier)

Format

Data frame contains the following columns:

BRKUSG3

factor; USGS-derived bedrock type map

CRPUSG3

numeric; percentage (0–1) of crop fields based on the LC6USGS

DEMSRT3

numeric; SRTM DEM elevations in meters

DINSRT3

numeric; diffuse insolation derived in SAGA GIS

EV1MOD3

numeric; MODIS time series (year 2011) Enhanced Vegetation Index PC1

EV2MOD3

numeric; MODIS time series (year 2011) Enhanced Vegetation Index PC2

GLCMER3

factor; GlobCov land cover values

INSSRT3

numeric; potential incoming solar radiation derived in SAGA GIS

LC6USG3

factor; USGS National Land Cover Database values (year 2006)

LNDUSG3

numeric; land mask for the Indiana state (0–1)

NL1DMS3

numeric; lights at night time series (1998–2010) images PC1

NL2DMS3

numeric; lights at night time series (1998–2010) images PC2

TD1MOD3

numeric; MODIS time series (year 2011) daytime Land Surface Temperature PC1

TD2MOD3

numeric; MODIS time series (year 2011) daytime Land Surface Temperature PC2

TN1MOD3

numeric; MODIS time series (year 2011) nighttime Land Surface Temperature PC1

TN2MOD3

numeric; MODIS time series (year 2011) nighttime Land Surface Temperature PC2

TRITOX3

numeric; kernel density of the Toxics Release Inventory (EPA) locations

TWISRT3

numeric; SAGA Topographic Wetness Index based on the SRTM DEM

SLPSRT3

numeric; slope in degrees based on the SRTM DEM

SMUSTS3

factor; mapping unit key from the General Soil Map STATSGO (at scale 1:250,000) obtained from the NRCS data gateway

SRGUSG3

factor; USGS-derived surface geology map

C01SUR3

numeric; percentage (0–1) of "Epiaqualfs"

C02SUR3

numeric; percentage (0–1) of "Endoaquolls"

C03SUR3

numeric; percentage (0–1) of "Hapludalfs"

C04SUR3

numeric; percentage (0–1) of "Eutrudepts"

C05SUR3

numeric; percentage (0–1) of "Haplosaprists"

C06SUR3

numeric; percentage (0–1) of "Hapludolls"

C07SUR3

numeric; percentage (0–1) of "Argiaquolls"

C09SUR3

numeric; percentage (0–1) of "Udorthents"

C10SUR3

numeric; percentage (0–1) of "Humaquepts"

C11SUR3

numeric; percentage (0–1) of "Fragiudalfs"

C12SUR3

numeric; percentage (0–1) of "Endoaqualfs"

C13SUR3

numeric; percentage (0–1) of "Fragiaqualfs"

C14SUR3

numeric; percentage (0–1) of "Dystrudepts"

C15SUR3

numeric; percentage (0–1) of "Fluvaquents"

C16SUR3

numeric; percentage (0–1) of "Hapludults"

C17SUR3

numeric; percentage (0–1) of "Argiudolls"

C18SUR3

numeric; percentage (0–1) of "Endoaquepts"

C19SUR3

numeric; percentage (0–1) of "Fragiudults"

C20SUR3

numeric; percentage (0–1) of "Glossaqualfs"

C21SUR3

numeric; percentage (0–1) of "Udipsamments"

C22SUR3

numeric; percentage (0–1) of "Paleudults"

C23SUR3

numeric; percentage (0–1) of "Paleudalfs"

C24SUR3

numeric; percentage (0–1) of "Udifluvents"

C25SUR3

numeric; percentage (0–1) of "Udarents"

C26SUR3

numeric; percentage (0–1) of "Epiaquepts"

C27SUR3

numeric; percentage (0–1) of "Psammaquents"

C31SUR3

numeric; percentage (0–1) of "Fragiaquults"

C33SUR3

numeric; percentage (0–1) of "Epiaquolls"

s1

numeric; x-coordinate (m) in the Albers equal-area projection system

s2

numeric; y-coordinate (m) in the Albers equal-area projection system

Details

These predictors were selected following the CLORPT concept of factors of soil formation, first introduced by Hans Jenny (1941), where C stands for climate, L for landform, O for organisms, R for relief, P for parent material and T for time. The study area spans four 1–degree cells i.e. from 38–40 degrees latitude and from -87 to -85 degrees longitude.
The percentage of crop fields and the land mask (0–100 percent) were derived using the area weighted averaging from the 30 m resolution land cover map (LC6USG3). The land mask (0–100 percent) shows percentage of water bodies, medium and high intensity build-up (urban) areas. To create a boolean mask map, consider using some treshold value (e.g. 0.6). SRTM DEM, MODIS and GlobCov are publicly available global data sets commonly used as standard global thematic layers for Digital Soil Mapping. These layers are avialable for download from the WorldGrids.org repository. Soil taxonomic classes (C**SUR3 (at scale 1:24,000), mapping units SMUSTS3), high resolution land cover data (LC6USG3), and geological maps (SRGUSG3 and BRKUSG3) are available only for this study area (USA National layers; obtained from the USGS Seamless data warehouse). Lights at night images have been obtained from the National Geophysical Data Centre. Density of the industrial facilities / pollutants (Toxics Release Inventory) were obtained from the U.S. National Library of Medicine. The first component from any time-series data usually shows long term mean value, while the higher order component show seasonal variations.

Note

Each listed layer follows a standard naming convention used by Worlgrids.org (the standard 8.3 filename convention with at most eight characters): first three letter are used for the variable type e.g. TD1 (daily surface temperature PC1); the next three letters represent the data source or collection method e.g. MOD (MODIS day time imagery); the 6th character is the effective scale e.g. 3 indicates the third standard scale (1/120 decimal degrees or about 1 km).
The great groups from the Soil Survey Geographic database (SSURGO) might show in some areas, i.e. for some soil taxa, artificial jumps (border areas between two soil surveys). SSURGO mapsheets are typically harmonized only after the reports are completed and such artifacts are common for regional soil surveys that are run independently one from the other.

Author(s)

Tomislav Hengl and Zamir Libohova

References

Savtchenko, A. and Ouzounov, D. and Ahmad, S. and Acker, J. and Leptoukh, G. and Koziana, J. and Nickless, D., (2004) Terra and Aqua MODIS products available from NASA GES DAAC. Advances in Space Research, 34(4): 710-714.

Soil Survey Staff, Natural Resources Conservation Service, United States Department of Agriculture. Soil Survey Geographic (SSURGO) Database for Indiata State. Available online at http://soildatamart.nrcs.usda.gov.

Soil Survey Staff, Natural Resources Conservation Service, United States Department of Agriculture. U.S. General Soil Map (STATSGO2). Available online at http://soildatamart.nrcs.usda.gov.

Xian, G., and Homer C., (2010) Updating the 2001 National Land Cover Database impervious surface products to 2006 using Landsat imagery change detection methods. Remote Sensing of Environment, 114, 1676-1686.

See Also

NASIS, NCSS, geochm

Examples


data(hoosier)
data(indiana)
border <- SpatialLines(list(Lines(list(Line(indiana)), "border")))
gridded(hoosier) <- ~s1+s2
## Not run: 
# overlay profiles and grids:
lst = c("C02SUR3","C04SUR3","C23SUR3")
col = grey(rev(seq(0.1,0.975,0.025)))
spplot(hoosier[lst], col.regions=col, sp.layout=list("sp.lines", border))
# write to a GIS format:
writeGDAL(hoosier["TD1MOD3"], "TD1MOD3.sdat", "SAGA")

## End(Not run)

[Package GSIF version 0.1-1 Index]