ilin {GSIF}	R Documentation

NSCD soil samples for Illinois and Indiana States

Description

A subset of the National Cooperative Soil Characterization Database (NSCD) covering the Illinois and Indiana States (5189 soil profiles). Analytical procedures and methods of soil preparation are identified by methods codes in the column headings of the standard NSSC data sheets and are described in Soil Survey Laboratory Information Manual No. 42.

Usage

data(ilin)

Format

Sites table. Contains the following columns:

SOURCEID

factor; unique label to help a user identify a particular site (user_site_id in the NSCD)

LATNAD83

numeric; latitude in decimal degrees on the NAD83 datum

LONNAD83

numeric; longitude in decimal degrees on the NAD83 datum

TIMESTRR

Date; the date on which this particular soil was described or sampled (observation_date in the NSCD)

TAXSUSDA

factor; correlated name for soil series (correlated_taxon_name in the NSCD)

TAXKUSDA

factor; taxon type — soil series or taxadjunct (correlated_taxon_kind in the NSCD)

TAXNUSDA

factor; correlated Keys to Soil Taxonomy taxon complete name e.g. "Sandy, mixed, thermic Ustic Torriorthent" (correlated_class_name in the NSCD)

Horizons table. Contains the following columns:

SOURCEID

factor; a short label to help a user identify a particular site (user_site_id in the NSCD)

LSQINT

integer; a layer sequence number 1 to N (layer_sequence in the NSCD)

HZDUSD

character; pedological designation of the horizon e.g. "Bt1" (hzn_desgn in the NSCD)

UHDICM

numeric; lower horizon depth from the surface in cm (hzn_bot in the NSCD)

LHDICM

numeric; upper horizon depth from the surface in cm (hzn_top in the NSCD)

CLYPPT

numeric; weight percentage of the clay particles (<0.0002 mm; clay_tot_psa in the NSCD)

SNDPPT

numeric; weight percentage of the silt particles (0.0002–0.05 mm; sand_tot_psa in the NSCD)

SLTPPT

numeric; weight percentage of the sand particles (0.05–2 mm; silt_tot_psa in the NSCD)

CRFMAS

numeric; weight percentage of coarse fragments (> 2 mm; wpg2 in the NSCD)

PHIHO5

numeric; pH index measured in water solution(ph_h2o in the NSCD)

PHIKCL

numeric; pH index measured in KCl solution (ph_kcl in the NSCD)

ORCDRC

numeric; soil organic carbon content in permille (see below)

BLDODR

bulk density oven dry (105 C) in kg per cubic meter (db_od in the NSCD)

Note

This data set is only a sample of the complete laboratory data. The original database contains over 30 thousand profiles. The variables listed above were selected for testing algorithms to produce soil property maps according to the GlobalSoilMap.net specifications. ORCDRC values are based on the c_tot and oc values in the NSCD. The total organic carbon in permille for alkaline soils (pH in water > 7) has been corrected by removing 12 percent of weight percentage of CaCO3.
Disclaimer: The assessment of the accuracy and applicability of the NSCD is strictly a USER RESPONSIBILITY. NRCS and NCSS take no responsibility for problems that arise from use of this database.

Author(s)

The NSCD data is maintained by the NRCS NSSC Soil Survey Lab, Lincoln (NE). This subset has been prepared for the purpose of testing various DSM methods by Tomislav Hengl <tom.hengl@wur.nl>. The processing steps are described in this script.

References

Soil Data Subcommittee, (1997) Soil Geographic Data Standard. Federal Geographic Data Committee, p. 286.

United States Department of Agriculture, Natural Resources Conservation Service, (2010) Keys to Soil Taxonomy, Eleventh Edition. Soil Survey Staff.

National Cooperative Soil Characterization Database (http://soils.usda.gov/survey/nscd/)

See Also

ilin.grids, geochm

Examples


data(ilin)
coordinates(ilin$sites) <- ~ LONNAD83 + LATNAD83
proj4string(ilin$sites) <- CRS("+proj=longlat +datum=NAD83")
## Not run: # obtain state borders:
require(maps)
states.m = map('state', plot=FALSE, fill=TRUE)
IDs <- sapply(strsplit(states.m$names, ":"), function(x) x[1])
require(maptools)
states <- as(map2SpatialPolygons(states.m, IDs=IDs), "SpatialLines")
# overlay and plot points and maps:
plot(spTransform(ilin$sites, CRS("+proj=longlat +datum=WGS84")), pch="+", cex=.6, col="black")
lines(states)

## End(Not run)

[Package GSIF version 0.2-0 Index]