isis {GSIF}	R Documentation

ISRIC Soil Information System

Description

ISRIC's collection of global soil monoliths that represent the main soil reference groups of the World Reference Base for Soil Resources (WRB). Includes some 950 monoliths (785 with coordinates) from over 70 countries with detailed soil profile and environmental data.

Usage

data(isis)

Format

The isis data set contains two data frames — sites and horizons. Sites table contains the following columns:

SOURCEID

factor; unique ISIS code

LONWGS84

numeric; longitude in decimal degrees on the WGS84 datum

LATWGS84

numeric; latitude in decimal degrees on the WGS84 datum

TIMESTRR

Date; the date on which this particular soil was described or sampled

TAXGWRB

factor; soil group based on the WRB classification system

TAXNUSDA

factor; Keys to Soil Taxonomy taxon name e.g. "Natraqualf"

BDRICM

numeric; depth to bedrock (R horizon) if observed

SOURCEDB

factor; source data base

Horizons table contains the following columns:

SOURCEID

factor; unique ISIS code

UHDICM

numeric; upper horizon depth from the surface in cm

LHDICM

numeric; lower horizon depth from the surface in cm

CRFVOL

numeric; volume percentage of coarse fragments (> 2 mm)

PHIHOX

numeric; pH index measured in water solution

PHIKCL

numeric; pH index measured in KCl solution

ORCDRC

numeric; soil organic carbon content in permilles

SNDPPT

numeric; weight percentage of the sand particles (0.05–2 mm)

SLTPPT

numeric; weight percentage of the silt particles (0.0002–0.05 mm)

CLYPPT

numeric; weight percentage of the clay particles (<0.0002 mm)

CEC

numeric; Cation Exchange Capacity in cmol+/kg

BLD

bulk density in tonnes per cubic-meter

Author(s)

ISRIC — World Soil Information

Examples


library(rgdal)
library(sp)

data(isis)
sites <- isis$sites
coordinates(sites) <- ~ LONWGS84 + LATWGS84
proj4string(sites) <- "+proj=longlat +datum=WGS84"
## Not run: 
## obtain country borders:
library(maps)
country.m = map('world', plot=FALSE, fill=TRUE)
IDs <- sapply(strsplit(country.m$names, ":"), function(x) x[1])
require(maptools)
country <- as(map2SpatialPolygons(country.m, IDs=IDs), "SpatialLines")
proj4string(country) = "+proj=longlat +datum=WGS84" 
## overlay and plot points and maps:
plot(country, col="darkgrey")
points(sites, pch=21, bg="red", cex=.6, col="black")

## End(Not run)

[Package GSIF version 0.4-2 Index]