landmask {GSIF}	R Documentation

Global coarse resolution land / soil mask maps

Description

Land mask showing the 1-degree cells (about 19 thousand in total) in the geographical coordinates, and the productive soils mask (areas with a positive Leaf Area Index at least once in the period 2002–2011). The land mask is based on the Global Self-consistent, Hierarchical, High-resolution Shoreline Database data (GSHHS 2.1), the productive soils mask on the MODIS Leaf Area Index monthtly product (MOD15A2), and the water mask is based on the MOD44W product. The map of the Keys to Soil Taxonomy soil suborders of the world at 20 km is based on the USDA-NRCS map of the global soil regions.

Usage

data(landmask)

Format

landmask data set is a data frame with the following columns:

mask

percent; land mask value

soilmask

boolean; soil mask value

watermask

percent; water mask value

Lon_it

indication of the longitude quadrant (W or E)

Lat_it

indication of the latitude quadrant (S or N)

cell_id

cell id code e.g. W79_N83

x

longitudes of the center of the grid nodes

y

latitudes of the center of the grid nodes

landmask20km data set is an object of class SpatialGridDataFrame with the following columns:

mask

percent; land mask value

suborder

factor; Keys to Soil Taxonomy suborder class e.g. Histels, Udolls, Calcids, ...

soilmask

factor; global soil mask map based on the land cover classes (see: SMKISR3)

Note

The land mask has been generated from the layer GSHHS_shp/h/GSHHS_h_L1.shp (level-1 boundaries).

References

Carroll, M., Townshend, J., DiMiceli, C., Noojipady, P., Sohlberg, R. (2009) A New Global Raster Water Mask at 250 Meter Resolution. International Journal of Digital Earth, 2(4).

Global Self-consistent, Hierarchical, High-resolution Shoreline Database (http://en.wikipedia.org/wiki/GSHHS)

USDA-NRCS Global Soil Regions Map (http://soils.usda.gov/use/worldsoils/)

Savtchenko, A., D. Ouzounov, S. Ahmad, J. Acker, G. Leptoukh, J. Koziana, and D. Nickless, (2004) Terra and Aqua MODIS products available from NASA GES DAAC. Advances in Space Research 34(4), 710-714.

Wessel, P., Smith, W.H.F., (1996) A Global Self-consistent, Hierarchical, High-resolution Shoreline Database. Journal of Geophysical Research, 101, 8741-8743.

See Also

rworldmap::rworldmapExamples, maps::map

Examples


library(rgdal)
library(sp)

data(landmask)
gridded(landmask) <- ~x+y
proj4string(landmask) <- "+proj=longlat +datum=WGS84"
## Not run:  ## plot maps:
require(maps)
country.m = map('world', plot=FALSE, fill=TRUE)
IDs <- sapply(strsplit(country.m$names, ":"), function(x) x[1])
require(maptools)
country <- as(map2SpatialPolygons(country.m, IDs=IDs), "SpatialLines")
spplot(landmask["mask"], col.regions="grey", sp.layout=list("sp.lines", country))
spplot(landmask["soilmask"], col.regions="grey", sp.layout=list("sp.lines", country))

## End(Not run)
## also available in the Robinson projection at 20 km grid:
data(landmask20km)
image(landmask20km[1])
summary(landmask20km$suborder)
summary(landmask20km$soilmask)

[Package GSIF version 0.4-1 Index]