make.3Dgrid {GSIF}	R Documentation

Methods to prepare 3D prediction locations

Description

Generates a list of objects of type "SpatialPixelsDataFrame" with longitude, latitude and altitude coordinates (these names are used by default for compatibility with the geosamples-class).

Usage


make.3Dgrid(obj, proj4string = get("ref_CRS", envir = GSIF.opts), 
    pixelsize = get("cellsize", envir = GSIF.opts)[1], 
    resample = "bilinear", NAflag = get("NAflag", envir = GSIF.opts), 
    stdepths = get("stdepths", envir = GSIF.opts), tmp.file = TRUE, 
    show.output.on.console = FALSE, ...)

Arguments

obj
object of class "SpatialPixelsDataFrame" or "RasterBrick"

pixelsize
grid cell size in decimal degrees (set by default at 0.0008333333 or 100 m around equator)

resample
character; resampling method to be passed the reprojection algorithm

NAflag
character; missing value flag

stdepths
numeric; list of standard depths

tmp.file
logical; specifies whether a temporary file name should be generated

show.output.on.console
logical; specifies whether to print out the progress

...
optional arguments that can be passed to the reprojetion algorithm

Details

The output is list of objects of class "SpatialPixelsDataFrame" where the number of elements in the list corresponds to the number of standard depths.

Note

If the input object is of class "SpatialPixelsDataFrame", the method by default uses FWTools (gdalwarp command) to resample grids, otherwise the raster::projectRaster command is passed. FWTools is a separate application and must be installed separately. This operation can be time consuming for large areas (e.g. continents).

Author(s)

Tomislav Hengl

See Also

spc, geosamples-class, plotKML::reproject

Examples


## Not run: # prepare prediction locations for Ebergotzen:
library(plotKML)
data(eberg_grid)
gridded(eberg_grid) <- ~x+y
proj4string(eberg_grid) <- CRS("+init=epsg:31467")
# convert to spatial components:
formulaString <- ~ PRMGEO6+DEMSRT6+TWISRT6+TIRAST6
eberg_spc <- spc(eberg_grid, formulaString)
# create 3D locations in the original coordinate system:
eberg_3Dxy <- sp3D(eberg_spc@predicted)
# wrapper function to create 3D locations in the default system:
eberg_3D <- make.3Dgrid(eberg_spc@predicted)
image(eberg_3D[[1]]["PC1"])

## End(Not run)

[Package GSIF version 0.2-1 Index]