merge {GSIF}	R Documentation

Merge multiple predictions

Description

Merges objects of class "SpatialPredictions" or "RasterBrickSimulations" and produces average predictions where the two objects overlap spatially. If the predictions are available at different resolutions, then it downscales all other grids to the smallest grid cell size using bicubic splines (for predictions) i.e. nearest neighbor algorithm (for simulations). Weigths can be passed via the RMSE.l argument, otherwise they will be estimated from validation slot (in the case of the "SpatialPredictions" class objects).

Usage


## S4 method for signature 'SpatialPredictions,SpatialPredictions'
merge(x, y, ..., 
          RMSE.l = NULL, silent = TRUE)

Arguments

x
object of class "SpatialPredictions" or "RasterBrickSimulations"

y
object of class "SpatialPredictions" or "RasterBrickSimulations"

...
additional objects of class "SpatialPredictions" or "RasterBrickSimulations"

RMSE.l
numeric; list of mean prediction errors for each object (these are used as weights during the averaging)

silent
logical; specifies whether to print out the progress and used RMSE's

Details

Merging of multiple predictions using weighted averaging is a heuristic approach to mapping. The main problem is that this method assumes that the predictions are completely independent (independent covariates, independent models), but this not might be the case and hence the predictions will be sub-optimal. Merging multiple predictions is however attractive for situations where the predictions do not have the same extent, so that a more global map can be used to fill in the gaps in locally produced predictions.

Value

Returns an object of type "SpatialPixelsDataFrame" or "RasterBrickSimulations" that contains only the merged values.

Author(s)

Tomislav Hengl and Gerard B.M. Heuvelink

References

Heuvelink, G.B.M., Bierkens, M.F.P. (1992) Combining soil maps with interpolations from point observations to predict quantitative soil properties. Geoderma 55(1-2): 1-15.

[Package GSIF version 0.3-1 Index]