riodoce {GSIF}	R Documentation

Soil samples for the Rio Doce basin

Description

Soil samples (456) collected for the Rio Doce basin in Brazil.

Usage

data(riodoce)

Format

Sites table. Contains the following columns:

SOURCEID

factor; unique name that identifies a particular site

LATWGS84

numeric; latitude in decimal degrees on the NAD83 datum

LONWGS84

numeric; longitude in decimal degrees on the NAD83 datum

TAXBRC

factor; original soil class based on the Brazilian Soil Classification System

Horizons table. Contains the following columns:

SOURCEID

factor; unique name that identifies a particular site

HRDLOC

character; local pedological designation of the horizon e.g. "Bt1"

UHDICM

numeric; upper horizon depth from the surface in cm

LHDICM

numeric; lower horizon depth from the surface in cm

PHIHO5

numeric; pH index measured in water solution

PHIKCL

numeric; pH index measured in KCl solution

ORCDRC

numeric; weight percentage in permilles of the total organic carbon

SNDPPT

numeric; weight percentage of the silt particles (0.0002–0.05 mm)

SLTPPT

numeric; weight percentage of the sand particles (0.05–2 mm)

CLYPPT

numeric; weight percentage of the clay particles (<0.0002 mm)

BSABCL

numeric; base saturation in percent

CEXBCL

numeric; Cation Exchange Capacity in cmol(+)/kg (summary)

BLDVOL

bulk density oven dry (105 C) in kg per cubic meter

AWAIMM

total available water capacity in mm

HORUID

factor; unique name that identifies a particular horizon

Author(s)

This soil profile data set has been prepared for the purpose of testing various soil mapping methods by Eliana de Souza and Tomislav Hengl. The processing steps are described in this script.

References

Benedetti, M.M. and Curi, N. and Sparovek, G. and de Carvalho Filho, A. and Silva, S.H.G., (2011) Updated Brazilian's Georeferenced Soil Database – an Improvement for International Scientific Information Exchanging. In: Principles, Application and Assessment in Soil Science (Edited by E. Burcu Özkaraova Güngör), InTech, chapter 16, pp. 307-330.

Filho, E.F. et al. (2011) Levantamento de solos e aptidão agrícola das terras da bacia do Rio Doce, Estado de Minas Gerais. Belo Horizonte/MG, 453 p. www.meioambiente.mg.gov.br

Cooper, M., Mendes. L.M.S., Silva, W.L.C, and Sparovek, G., (2005) A National Soil Profile Database for Brazil Available to International Scientists. Soil Science Society of America Journal, 69:649-652.

See Also

riodoce.grids

Examples


## Not run: 
require(aqp)
data(riodoce)
# Promote to SoilProfileCollection
riodoce <- join(riodoce$horizons, riodoce$sites, type='inner')
# specify point IDs and depths:
depths(riodoce) <- SOURCEID ~ UHDICM + LHDICM
# extract site data
site(riodoce) <- ~ LONWGS84 + LATWGS84 + TAXBRC  
# generate SpatialPoints
coordinates(riodoce) <- ~ LONWGS84 + LATWGS84
proj4string(riodoce) <- "+proj=latlong +datum=WGS84"
str(riodoce)
plot(riodoce@sp, pch="+")

## End(Not run)

[Package GSIF version 0.2-0 Index]