soil.legends {GSIF}	R Documentation

Standard color palettes for soil properties and classes

Description

Standard color palettes for soil properties and classes that can be used to display global soil data.

Usage

data(soil.legends)

Format

Contains a list of color palettes (data frames with class names / break points, and cumulative probabilities) for:

ORCDRC

numeric; soil organic carbon content in permille

PHIHO5

numeric; pH index measured in water solution

PHIHO5

numeric; pH index measured in KCl solution

BLD

numeric; bulk density in kg per cubic meter

CEC

numeric; Cation Exchange Capacity

SNDPPT

numeric; weight percentage of the sand particles (0.05–2 mm)

SLTPPT

numeric; weight percentage of the silt particles (0.0002–0.05 mm)

CLYPPT

numeric; weight percentage of the clay particles (<0.0002 mm)

CRFVOL

numeric; volumetric percentage of coarse fragments (>2 mm)

TAXGWRB

factor; World Reference base groups

TAXOUSDA

factor; Keys to Soil Taxonomy suborders

Note

Breaks for continuous soil properties were determined using the quantiles function and by visually inspecting the histograms to maximize the contrast in output maps. Based on a compilation of global soil profile data (http://soilprofiles.org).

Author(s)

Tomislav Hengl

References

Global Soil Information Facilities (http://gsif.isric.org)

Examples


data(soil.legends)
pal <- soil.legends$ORCDRC$COLOR
names(pal) <- signif((soil.legends$ORCDRC$MAX + 
    soil.legends$ORCDRC$MIN)/2, 3)
pal
## make SAGA GIS palette:
makeSAGAlegend(x=as.factor(names(pal)), col_pal=pal, 
    filename="ORCDRC.txt")

[Package GSIF version 0.3-6 Index]