spc {GSIF}	R Documentation

Derive Soil Predictive Components

Description

Derives Soil Predictive Components for a given set of covariates. It wraps the stats::prcomp method and predicts a list principal components using the same grid system.

Usage


spc( )

Arguments

obj
object of class "SpatialPixelsDataFrame"; must contain at least two grids

Details

The output is a "SpatialPixelsDataFrame" object with generic names PC1,...,PCp, where p is the total number of input grids.

Value

spc returns an object of type "SpatialComponents".

Note

This method assumes that the input covariates are cross-correlated and hence their overlap can be reduced. This method scales input variables by default and replaces all missing values with 0 values, to reduce loss of data due to missing pixels. This operation can be time consuming for large grids.

Author(s)

Tomislav Hengl

See Also

stats::prcomp, SpatialComponents-class

Examples


## Not run: # load data:
library(plotKML)
pal = rev(rainbow(65)[1:48])
data(eberg_grid)
gridded(eberg_grid) <- ~x+y
proj4string(eberg_grid) <- CRS("+init=epsg:31467")
eberg_spc <- spc(eberg_grid, as.formula("~ PRMGEO6+DEMSRT6+TWISRT6+TIRAST6"))
names(eberg_spc@sp) # 11 components on the end;
rd = range(eberg_spc@sp@data[,1], na.rm=TRUE)
spplot(eberg_spc@sp[1:4], at=seq(rd[1], rd[2], length.out=48), col.regions=pal)

## End(Not run)
{methods}

[Package GSIF version 0.2-0 Index]