summary-methods {GSIF}	R Documentation

Summarize an object of class "SpatialPredictions"

Description

Derives a statistical summary for an object of class "SpatialPredictions".

Usage


## S4 method for signature 'SpatialPredictions'
summary(object)

Arguments

object
object of class "SpatialPredictions"

Details

The function creates a summary table with standard column names. These tell us what is the summary accuracy of the spatial predictions and what are the effective bytes of information produced.

Value

The summary returns a data.frame with the following columns:

"variable"

variable name

"minium"

lowest value observed

"maximum"

largest value observed

"npoints"

number of observations

"area"

lowest value observed

"area.units"

area units either square-m or square-arcdegrees

"covariates"

list of covariates used

"family"

GLM family (if applicable)

"RMSE"

RMSE derived using cross-validation

"tvar"

variance percent explained by the model using the cross-validation

"npixels"

total number of produced pixels

"breaks"

breaks based on the half RMSE

"bonds"

lower and upper boundaries for effective classes

"Bytes"

effective bytes produced (see Hengl et al (2012) for more details)

"compress"

compression algorithm used

Author(s)

Tomislav Hengl

References

Hengl, T., Nikolic, M., MacMillan, R.A., (2013) Mapping efficiency and information content. International Journal of Applied Earth Observation and Geoinformation, special issue Spatial Statistics Conference, 22: 127–138.

See Also

plotKML::SpatialPredictions-class

Examples


## load observations:
library(sp)
library(rgdal)
library(gstat)
demo(meuse, echo=FALSE)
## fit a model:
omm <- fit.gstatModel(meuse, om~dist, 
  fit.family=gaussian(link="log"), meuse.grid)
show(omm@regModel)
## produce SpatialPredictions:
om.rk <- predict(omm, predictionLocations = meuse.grid)
x = summary(om.rk)
str(x)

[Package GSIF version 0.4-3 Index]