tile {GSIF}	R Documentation

Tiles (subsets or clips) a spatial object to regular blocks

Description

Tiles objects of class "Spatial*" or "RasterLayer*" into regular blocks.

Usage


## S4 method for signature 'SpatialPointsDataFrame'
tile(x, y, block.x, ...)
## S4 method for signature 'SpatialPixelsDataFrame'
tile(x, y, block.x, ...)
## S4 method for signature 'SpatialPolygonsDataFrame'
tile(x, y, block.x, tmp.file = FALSE, 
    program, show.output.on.console = FALSE, ...)
## S4 method for signature 'SpatialLinesDataFrame'
tile(x, y, block.x, tmp.file = FALSE, 
    program, show.output.on.console = FALSE, ...)
## S4 method for signature 'RasterLayer'
tile(x, y, block.x, tmp.file = FALSE, 
    program, show.output.on.console = FALSE, ...)

Arguments

x
object of class "Spatial*" "RasterLayer"

y
list of "SpatialPolygons"; if missing will be derived based on block.x

block.x
numeric; size of block in meters or corresponding mapping units

tmp.file
logical; specifies whether to generate a temporary file

program
character; location of the auxiliary program in the system

show.output.on.console
logical; specifies whether to print the progress of a function

...
optional arguments that can be passed to the getSpatialTiles

Details

When working with objects of type "SpatialLinesDataFrame", "SpatialPolygonsDataFrame" and or "RasterLayer", the function looks for FWTools binary files ogr2ogr and gdalwarp. FWTools is a separate program and must be installed separately.

Value

Returns a list of objects of the same class as the input object.

Author(s)

Tomislav Hengl

See Also

getSpatialTiles

Examples


## spatial pixels:
library(sp)
data(meuse.grid)
gridded(meuse.grid) <- ~x+y
tl <- getSpatialTiles(meuse.grid, block.x=1000)
image(meuse.grid)
lines(as(tl, "SpatialLines"))
# all at once:
pix.lst <- tile(meuse.grid, block.x=1000)
## Not run: ## lines:
library(plotKML)
data(eberg_contours)
line.lst <- tile(eberg_contours, block.x=5000)
spplot(line.lst[[1]][2])
## polygons:
data(eberg_zones)
## this one requires ogr2ogr function:
pol.lst <- tile(eberg_zones, block.x=5000)
spplot(pol.lst[[1]][2])

## End(Not run)

[Package GSIF version 0.3-0 Index]