SpatialComponents-class {GSIF}	R Documentation

A class for gridded components derived using the spc method

Description

A class containing a list of gridded components and results of principal component analysis.

Slots

predicted:

object of class "SpatialPixelsDataFrame"; predicted values for components

pca:

object of class "list"; output objects from the stats::prcomp process — contains objects: 'stdev', 'rotation', 'center' and 'scale'

Author(s)

Tomislav Hengl

See Also

spc

[Package GSIF version 0.3-0 Index]