afsp {GSIF}	R Documentation

Africa Soil Profiles Database

Description

A subset of the Africa Soil Profiles Database (AFSP) that contains over 12,000 geo-referenced legacy soil profile records for 37 countries.

Usage

data(afsp)

Format

Sites table. Contains the following columns:

SOURCEID

factor; unique label to help a user identify a particular site (user_site_id in the NSCD)

RLBISRIC

factor; reliability class according to the ISRIC quality standards

LONWGS84

numeric; longitude in decimal degrees on the WGS84 datum

LATWGS84

numeric; latitude in decimal degrees on the WGS84 datum

STDXYZ

numeric; profile location accuracy in decimal degrees

TIMESTRR

character; the date on which this particular soil was described or sampled

TAXNWRB

factor; full soil classification name based on the WRB classification system

TAXGWRB

factor; abbreviated soil group based on the WRB classification system

TAXNUSDA

factor; Keys to Soil Taxonomy taxon name e.g. "Plinthic Udoxic Dystropept"

Horizons table. Contains the following columns:

SOURCEID

factor; a short label to help a user identify a particular site (ProfileID in the AFSP)

LSQINT

integer; a layer sequence number 1 to N (LayerNr in the AFSP)

UHDICM

numeric; upper horizon depth from the surface in cm (UpDpth in the AFSP)

LHDICM

numeric; lower horizon depth from the surface in cm (LowDpth in the AFSP)

MCOMNS

factor; Munsell color moist (ColorM in the NSCD)

CRFVOL

numeric; volume percentage of coarse fragments (> 2 mm; CfPc in the AFSP)

SNDPPT

numeric; weight percentage of the silt particles (0.0002–0.05 mm; Sand in the AFSP)

SLTPPT

numeric; weight percentage of the sand particles (0.05–2 mm; Silt in the AFSP)

CLYPPT

numeric; weight percentage of the clay particles (<0.0002 mm; Clay in the AFSP)

BLD

bulk density in kg per cubic meter (BlkDens in the AFSP)

PHIHO5

numeric; pH index measured in water solution(PHH2O in the AFSP)

PHIKCL

numeric; pH index measured in KCl solution (PHKCl in the AFSP)

ORCDRC

numeric; soil organic carbon content in permille (OrgC in the AFSP))

Author(s)

The Africa Soil Profiles Database have been prepared by Johan Leenaars <johan.leenaars@wur.nl>. This is a subset of the original database that can be downloaded via www.isric.org. The processing steps used to prepare this data are described in this script.

References

Leenaars, J.G.B. (2012) Africa Soil Profiles Database, Version 1.0. A compilation of geo-referenced and standardized legacy soil profile data for Sub Saharan Africa (with dataset). ISRIC report 2012/03. Africa Soil Information Service (AfSIS) project and ISRIC – World Soil Information, Wageningen, the Netherlands.

Africa Soil Information Service (http://africasoils.net)

See Also

mw.grids

Examples


data(afsp)
coordinates(afsp$sites) <- ~ LONWGS84 + LATWGS84
proj4string(afsp$sites) <- CRS("+proj=longlat +datum=WGS84")
## Not run: # obtain state borders:
require(maps)
country.m = map('world', plot=FALSE, fill=TRUE)
IDs <- sapply(strsplit(country.m$names, ":"), function(x) x[1])
require(maptools)
country <- as(map2SpatialPolygons(country.m, IDs=IDs), "SpatialLines")
# overlay and plot points and maps:
plot(afsp$sites, pch="+", cex=.6, col="black")
lines(country)

## End(Not run)

[Package GSIF version 0.2-0 Index]