afsp {GSIF}	R Documentation

Africa Soil Profiles Database

Description

A subset of the Africa Soil Profiles Database (AFSP) that contains over 12,000 geo-referenced legacy soil profile records for 37 countries.

Usage

data(afsp)

Format

The afsp data set contains two data frames — sites and horizons. Sites table contains the following columns:

SOURCEID

factor; unique label to help a user identify a particular site (ProfileID in the AFSP)

RLBISRIC

factor; reliability class according to the ISRIC quality standards (Reliab in the AFSP)

LONWGS84

numeric; longitude in decimal degrees on the WGS84 datum (X_LonDD in the AFSP)

LATWGS84

numeric; latitude in decimal degrees on the WGS84 datum (Y_LatDD in the AFSP)

STDXYZ

numeric; profile location accuracy in decimal degrees (XYAccur in the AFSP)

TIMESTRR

character; the date on which this particular soil was described or sampled (T_Year in the AFSP)

TAXNWRB

factor; full soil classification name based on the WRB classification system (ProfileID in the AFSP)

WRB06

factor; abbreviated soil group based on the WRB classification system (WRB06rg in the AFSP)

TAXNUSDA

factor; Keys to Soil Taxonomy taxon name e.g. "Plinthic Udoxic Dystropept" (USDA in the AFSP)

Horizons table contains the following columns:

SOURCEID

factor; a short label to help a user identify a particular site (ProfileID in the AFSP)

LSQINT

integer; a layer sequence number 1 to N (LayerNr in the AFSP)

UHDICM

numeric; upper horizon depth from the surface in cm (UpDpth in the AFSP)

LHDICM

numeric; lower horizon depth from the surface in cm (LowDpth in the AFSP)

MCOMNS

factor; Munsell color moist (ColorM in the NSCD)

CRFVOL

numeric; volume percentage of coarse fragments (> 2 mm; CfPc in the AFSP)

SNDPPT

numeric; weight percentage of the silt particles (0.0002–0.05 mm; Sand in the AFSP)

SLTPPT

numeric; weight percentage of the sand particles (0.05–2 mm; Silt in the AFSP)

CLYPPT

numeric; weight percentage of the clay particles (<0.0002 mm; Clay in the AFSP)

BLD

bulk density in kg per cubic meter (BlkDens in the AFSP)

PHIHO5

numeric; pH index measured in water solution(PHH2O in the AFSP)

PHIKCL

numeric; pH index measured in KCl solution (PHKCl in the AFSP)

ORCDRC

numeric; soil organic carbon content in permille (OrgC in the AFSP)

The mw.grids data frame contains a list of soil covariates prepared for Malawi (94,215 grid nodes):

DEMSRT3

numeric; SRTM DEM

EV1MOD3

numeric; first principal component of the monthly MODIS EVI time series data

EV2MOD3

numeric; second principal component of the monthly MODIS EVI time series data

GLCESA3

factor; general land cover class based on the GlobCov v2.2

INSSRT3

numeric; potential incoming solar radiation derived in SAGA GIS

SLPSRT3

numeric; slope in degrees based on the SRTM DEM

TAXWRB3

factor; dominant soil class (group) following the World Reference Base soil classification system derived from the general soil polygon map of Malawi

TD1MOD3

numeric; first principal component of the 8-day MODIS day-time LST time series data

TN1MOD3

numeric; first principal component of the 8-day MODIS night-time LST time series data

TWISRT2

numeric; SAGA Topographic Wetness Index based on the SRTM DEM

x

numeric; x-coordinate in the UTM 36S coordiante system with WGS84 ellipsoid

y

numeric; y-coordinate in the UTM 36S coordiante system with WGS84 ellipsoid

Note

The landscape of Malawi is a series of dissected plateaus and mountainous areas along the Rift Valley i.e. lake Malawi. The size of the country is 118,000 square kilometers (the size of the rectangular block is 349 by 860 km). Dominat soil type map (TAXWRB3) was derived from the 1:600k scale soils map of Malawi and rasterized to 1 km resolution. Listed gridded layers follow a standard naming convention used by WorlGrids.org (the standard 8.3 filename convention with at most eight characters): first three letter are used for the variable type e.g. DEM (digital elevation model); the next three letters represent the data source or collection method e.g. SRT (SRTM); the 6th character is the effective scale e.g. 3 indicates the 2nd standard scale i.e. 1/120 decimal degrees (about 1 km).

Author(s)

The Africa Soil Profiles Database have been prepared by Johan Leenaars <johan.leenaars@wur.nl>. This is a subset of the original database that can be downloaded via www.isric.org. The processing steps used to prepare this data are described in this script.

References

Leenaars, J.G.B. (2012) Africa Soil Profiles Database, Version 1.0. A compilation of geo-referenced and standardized legacy soil profile data for Sub Saharan Africa (with dataset). ISRIC report 2012/03. Africa Soil Information Service (AfSIS) project and ISRIC — World Soil Information, Wageningen, the Netherlands.

Ministry of Agriculture Malawi / FAO, (1992) Soils map of Malawi. Land Resources Evaluation Project /MLW/85/011. Land Resources Conservation Board, Lilongwe.

Africa Soil Information Service (http://africasoils.net)

Examples


## Not run: # load data and convert to SPC:
data(afsp)
sites <- afsp$sites
coordinates(sites) <- ~ LONWGS84 + LATWGS84
proj4string(sites) <- "+proj=latlong +datum=WGS84"
# obtain country borders:
library(maps)
country.m = map('world', plot=FALSE, fill=TRUE)
IDs <- sapply(strsplit(country.m$names, ":"), function(x) x[1])
require(maptools)
country <- as(map2SpatialPolygons(country.m, IDs=IDs), "SpatialLines")
# overlay and plot points and maps:
plot(country, col="darkgrey", xlim=c(-25.3,57.8), ylim=c(-34.8, 37.4))
points(sites, pch=21, bg="white", cex=.6, col="black")
# load gridded data for Malawi:
download.file("http://worldgrids.org/rda/mw.grids.rda", "mw.grids.rda")
load("mw.grids.rda")
gridded(mw.grids) <- ~x+y
proj4string(mw.grids) <- CRS(utm.csy)
# plot maps:
col = grey(rev(seq(0.1,0.975,0.025)))
names(mw.grids)
spplot(mw.grids[1], col.regions=col)
# write to a GIS format:
writeGDAL(mw.grids["EV1MOD3"], "EV1MOD3.tif", "GTiff", mvFlag=-99999)

## End(Not run)

[Package GSIF version 0.2-4 Index]