afsp {GSIF}	R Documentation

Africa Soil Profiles Database

Description

A subset of the Africa Soil Profiles Database (AFSP) that contains over 17,000 geo-referenced legacy soil profile records for 37 countries.

Usage

data(afsp)

Format

The afsp data set contains two data frames — sites and horizons. Sites table contains the following columns:

SOURCEID

factor; unique label to help a user identify a particular site (ProfileID in the AFSP)

RLBISRIC

factor; reliability class according to the ISRIC quality standards (Reliab in the AFSP)

LONWGS84

numeric; longitude in decimal degrees on the WGS84 datum (X_LonDD in the AFSP)

LATWGS84

numeric; latitude in decimal degrees on the WGS84 datum (Y_LatDD in the AFSP)

STDXYZ

numeric; profile location accuracy in decimal degrees (XYAccur in the AFSP)

TIMESTRR

character; the date on which this particular soil was described or sampled (T_Year in the AFSP)

TAXNWRB

factor; full soil classification name based on the WRB classification system (ProfileID in the AFSP)

TAXNFAO

factor; full soil classification code based on the FAO 1988 classification system (FAO88 in the AFSP)

TAXGWRB

factor; abbreviated soil group based on the WRB classification system (WRB06rg in the AFSP)

TAXNUSDA

factor; Keys to Soil Taxonomy taxon name e.g. "Plinthic Udoxic Dystropept" (USDA in the AFSP)

DRAINFAO

factor; drainage class based on the FAO guidelines for soil description: E (excessively drained), S (somewhat excessively drained), W (well drained), M (moderately well drained), I (somewhat poorly drained) and V (very poorly drained)

Horizons table contains the following columns:

SOURCEID

factor; a short label to help a user identify a particular site (ProfileID in the AFSP)

LSQINT

integer; a layer sequence number 1 to N (LayerNr in the AFSP)

UHDICM

numeric; upper horizon depth from the surface in cm (UpDpth in the AFSP)

LHDICM

numeric; lower horizon depth from the surface in cm (LowDpth in the AFSP)

MCOMNS

factor; Munsell color moist (ColorM in the NSCD)

CRFVOL

numeric; volume percentage of coarse fragments (> 2 mm; CfPc in the AFSP)

SNDPPT

numeric; weight percentage of the sand particles (0.0002–0.05 mm; Sand in the AFSP)

SLTPPT

numeric; weight percentage of the silt particles (0.05–2 mm; Silt in the AFSP)

CLYPPT

numeric; weight percentage of the clay particles (<0.0002 mm; Clay in the AFSP)

BLD

bulk density in tonnes per cubic-meter (BlkDens in the AFSP)

PHIHOX

numeric; pH index measured in water solution (PHH2O in the AFSP)

PHIKCL

numeric; pH index measured in KCl solution (PHKCl in the AFSP)

ECN

numeric; electrical conductivity (EC in the AFSP)

ORCDRC

numeric; soil organic carbon content in permilles (OrgC in the AFSP)

CEC

numeric; Cation Exchange Capacity in cmol+/kg (CecSoil in the AFSP)

EACKCL

numeric; exchangable acidity (ExAcid in the AFSP)

The afgrid20 "SpatialPixelsDataFrame" object contains a list of soil covariates prepared for Africa at 20 km resolution (45,997 grid nodes) obtained from WorldGrids.org:

LMTGSH0

integer; land mask based on GSHHS

SMKMOD0

integer; mask map indicating soil productive areas

DEMSRE0

integer; global relief model based on SRTM 30+ and ETOPO DEM at 1/120 arc degrees

SLPSRT0

integer; slope map in percent derived using the DEMSRE

TWISRE0

numeric; SAGA Topographic Wetness Index based on the SRTM DEM

G01ESA0

integer; percentage of land cover class 1 based on the GlobCov v2.2

...

other land cover classes

IFLGRE0

integer; intact forest landscapes based on Greenpeace

L3POBI0

factor; physiographic landform units (SCALA project)

PREGSM0

numeric; mean monthly precipitation (annual)

TDHMOD0

integer; maximum value of the 8-day MODIS day-time LST time series data

TDMMOD0

integer; mean value the 8-day MODIS day-time LST time series data

TDSMOD0

integer; standard deviation of the 8-day MODIS day-time LST time series data

TNMMOD0

integer; mean value the 8-day MODIS night-time LST time series data

TNSMOD0

integer; standard deviation of the 8-day MODIS night-time LST time series data

EVMMOD0

numeric; mean value of the monthly MODIS EVI time series data

EVSMOD0

integer; standard deviation of the monthly MODIS EVI time series data

SGEUSG0

factor; parent material class based on the USGS Surface Geology map

Note

The soil covariate layers were selected to represent the standard soil forming factors: climate, organism, relief, and parent materials (CLORPT). Listed gridded layers follow a standard naming convention used by WorlGrids.org (the standard 8.3 filename convention with at most eight characters): first three letter are used for the variable type e.g. DEM (digital elevation model); the next three letters represent the data source or collection method e.g. SRE (SRTM/ETOPO); the 6th character is the effective scale e.g. 0 indicates grid size of i.e. 1/5 decimal degrees (about 20 km).

Author(s)

The Africa Soil Profiles Database have been prepared by Johan Leenaars <johan.leenaars@wur.nl>. This is a subset of the original database that can be downloaded via www.isric.org. The processing steps used to prepare this data are described in this script.

References

Leenaars, J.G.B. (2014) Africa Soil Profiles Database, Version 1.2. A compilation of geo-referenced and standardized legacy soil profile data for Sub Saharan Africa (with dataset). ISRIC report 2012/03. Africa Soil Information Service (AfSIS) project and ISRIC — World Soil Information, Wageningen, the Netherlands.

Africa Soil Information Service (http://africasoils.net)

Examples


## Not run: 
library(rgdal)
library(aqp)
library(sp)

data(afsp)
sites <- afsp$sites
coordinates(sites) <- ~ LONWGS84 + LATWGS84
proj4string(sites) <- "+proj=longlat +datum=WGS84"
## obtain country borders:
library(maps)
country.m = map('world', plot=FALSE, fill=TRUE)
IDs <- sapply(strsplit(country.m$names, ":"), function(x) x[1])
require(maptools)
country <- as(map2SpatialPolygons(country.m, IDs=IDs), "SpatialLines")
proj4string(country) = "+proj=longlat +datum=WGS84" 
## overlay and plot points and maps:
plot(country, col="darkgrey", xlim=c(-25.3,57.8), ylim=c(-34.8, 37.4))
points(sites, pch=21, bg="white", cex=.6, col="black")
## load gridded data for Africa:
con <- url("http://gsif.isric.org/lib/exe/fetch.php?media=afgrid20.rda")
load(con)
## plot maps:
col = grey(rev(seq(0.1,0.975,0.025)))
names(afgrid20)
lns = list("sp.lines", spTransform(country, afgrid20@proj4string))
spplot(afgrid20["TDMMOD0a"], col.regions=col, sp.layout=lns)
## write to a GIS format:
writeGDAL(afgrid20["TDMMOD0a"], "TDMMOD0a.tif", "GTiff", mvFlag=-99999)

## End(Not run)

[Package GSIF version 0.4-2 Index]