as.geosamples {GSIF}	R Documentation

Converts an object of class "SoilProfileCollection" to a standardized data frame

Description

Converts an object of class "SoilProfileCollection" to an object of class "geosamples" with all measurements broken separately as individual records. Geosamples are standardized spatially and temporally referenced samples from the Earth's surface.

Usage


as.geosamples(obj, registry = as.character(NA), sample.area = 1, mxd = 2, dtime = 3600)

Arguments

obj
object of class "SoilProfileCollection"

registry
URI specifying the metadata registry

sample.area
standard sample area in square meters

mxd
maximum depth of interest in meters

dtime
standard temporal support in seconds

Details

Returns an object of type "geosamples". Many columns required by the "geosamples" class might be not available and will result in NA values. To ensure compatibility, when building an object of type "SoilProfilesCollection", use some standard naming convention to attach attributes:

"locationError"

can be used to attach location errors in meters to each spatial location

"measurementError"

can be used to attach specific measurement errors to each measurement in both site and horizons table

"IGSN"

can be used to attach the unique identifier (International Geo Sample Number) to each specific measurement

Author(s)

Tomislav Hengl and Hannes I. Reuter

See Also

geosamples-class, as.data.frame, aqp::SoilProfileCollection

Examples


require(aqp)
# sample profile from Nigeria:
lon = 3.90; lat = 7.50; time = as.POSIXct("1978", format="%Y"); id = "ISRIC:NG0017"; TAXNFAO8 = "LXp" 
top = c(0, 18, 36, 65, 87, 127) 
bottom = c(18, 36, 65, 87, 127, 181)
ORCDRC = c(18.4, 4.4, 3.6, 3.6, 3.2, 1.2)
methodid = c("TAXNFAO8", "ORCDRC")
description = c("FAO 1988 classification system group", "Method of Walkley-Black (Org. matter = Org. C x 1.72)")
units = c("FAO 1988 classes", "permille")
detectionLimit = c(as.character(NA), "0.1")
# prepare a SoilProfileCollection:
prof1 <- join(data.frame(id, top, bottom, ORCDRC), data.frame(id, lon, lat, time, TAXNFAO8), type='inner')
depths(prof1) <- id ~ top + bottom
site(prof1) <- ~ lon + lat + time + TAXNFAO8 
coordinates(prof1) <- ~ lon + lat + time
proj4string(prof1) <- CRS("+proj=longlat +datum=WGS84")
# add measurement errors:
attr(prof1@horizons$ORCDRC, "measurementError") <- c(1.5, 0.5, 0.5, 0.5, 0.5, 0.5)
attr(prof1@sp@coords, "locationError") <- 1500
# add the metadata:
prof1@metadata <- data.frame(methodid, description, units, detectionLimit)
# convert to geosamples:
x <- as.geosamples(prof1)
show(x)
# print only the sampled values of ORCDRC:
ORCDRC <- subset(x, "ORCDRC")

[Package GSIF version 0.2-0 Index]