fit.gstatModel-methods {GSIF}	R Documentation

Methods to fit a regression-kriging model

Description

Tries to automatically fit a 2D or 3D regression-kriging model for a given set of points and covariates (object of type "SpatialPixelsDataFrame"). It first runs GLM following the formulaString model, then fits variogram for residuals usign the fit.variogram method from the package gstat. Creates an output object of class gstatModel-class.

Usage

 fit.gstatModel(observations, formulaString, covariates, ... )

Arguments

observations
object of type "SpatialPointsDataFrame"

formulaString
object of type "formula" or a character

covariates
object of type "SpatialPixelsDataFrame"

...
other optional arguments that can be passed to glm and/or fit.variogram

Details

Optional arguments are vgmFun — variogram model type passed from gstat; family — character string defyning the GLM family (for more info see stats::glm); stepwise — specifies whether to run step-wise regression on top of GLM to get an optimal subset of predictors. The method by default assumes that the target variable follows a normal distribution family = gaussian. Other possible families are:

normal distribution

family = gaussian (default setting)

log-normal distribution

family = gaussian(log)

binomial variable

family = binomial(logit)

variable following a poisson distribution

family = poisson(log)

Author(s)

Tomislav Hengl and Gerard B.M. Heuvelink

References

Minasny, B., McBratney, A.B. (2007) Spatial prediction of soil properties using EBLUP with the Matérn covariance function. Geoderma, 140(4): 324-336.

Hengl, T. (2009) A Practical Guide to Geostatistical Mapping, 2nd Edt. University of Amsterdam, www.lulu.com, 291 p.

See Also

gstatModel-class, stats::glm, gstat::fit.variogram

Examples


data(meuse)
coordinates(meuse) <- ~x+y
# load grids:
data(meuse.grid)
coordinates(meuse.grid) <- ~x+y
gridded(meuse.grid) <- TRUE
# fit a model:
omm <- fit.gstatModel(observations = meuse, om~dist, covariates = meuse.grid,
  family = gaussian(log))
omm@regModel
om.sp <- SpatialPointsDataFrame(omm@sp, data = omm@regModel$model)
## Not run: # plot a variogram:
plot(variogram(om~1, om.sp))

## End(Not run)

[Package GSIF version 0.2-1 Index]