ilin {GSIF}	R Documentation

NSCD soil samples for Illinois and Indiana States

Description

A subset of the National Cooperative Soil Characterization Database (NSCD) covering the Illinois and Indiana States (5193 soil profiles). Analytical procedures and methods of soil preparation are identified by methods codes in the column headings of the standard NSSC data sheets and are described in Soil Survey Laboratory Information Manual No. 42.

Usage

data(ilin)

Format

Sites table. Contains the following columns:

site_key

factor; unique name that identifies a particular site

user_site_id

character; a short label to help a user identify a particular site

LAT

numeric; latitude in decimal degrees on the NAD83 datum

LON

numeric; longitude in decimal degrees on the NAD83 datum

horizontal_datum_name

character; horizontal datum name (if different from NAD83)

observation_date

character; the date on which this particular soil was described or sampled

sampled_taxon_name

character; KST soil series or taxadjunct name e.g. "Paxton"

correlated_taxon_name

character; correlated name for soil series

correlated_taxon_kind

character; taxon type — soil series or taxadjunct

correlated_class_name

character; correlated Keys to Soil Taxonomy taxon complete name e.g. "Sandy, mixed, thermic Ustic Torriorthent"

SSL_class_name

characted; original Keys to Soil Taxonomy taxon designation

Horizons table. Contains the following columns:

site_key

factor; unique name that identifies a particular site

natural_key

character; an integer number that uniquely identifies a particular horizon

layer_sequence

integer; a layer sequence number 1 to N

hzn_desgn

character; pedological designation of the horizon e.g. "Bt1"

hzn_bot

numeric; lower horizon depth from the surface in cm

hzn_top

numeric; upper horizon depth from the surface in cm

hzn_vert_subdvn

integer; an integer number that identifies horizon subdivision e.g. "Bt1" and "Bt2" both receive number "2"

clay_tot_psa

numeric; weight percentage of the clay particles (<0.0002 mm)

sand_tot_psa

numeric; weight percentage of the silt particles (0.0002–0.05 mm)

silt_tot_psa

numeric; weight percentage of the sand particles (0.05–2 mm)

wpg2

numeric; weight percentage of coarse fragments (> 2 mm)

caco3

numeric; weight percentage of carbonates in the soil

oc

numeric; weight percentage of organic carbon

c_tot

numeric; weight percentage of the total organic and inorganic carbon

base_sum

numeric; base saturation in percent

cec_sum

numeric; Cation Exchange Capacity in cmol(+)/kg (summary)

cec_nh4

numeric; Cation Exchange Capacity in cmol(+)/kg (NH4OAc)

ph_h2o

numeric; pH index measured in water solution

ph_kcl

numeric; pH index measured in KCl solution

db_od

bulk density oven dry (105 C) in kg per cubic meter

layer_ID

factor; unique name that identifies a particular horizon

Note

This data set is only a sample of the complete laboratory data. The original database contains over 30 thousand profiles. The variables listed above were selected for testing algorithms to produce soil property maps according to the GlobalSoilMap.net specifications. The total organic carbon for alkaline soils (pH in water > 7) needs to be corrected e.g. by removing 12 percent of weight percentage of CaCO3.
Disclaimer: The assessment of the accuracy and applicability of the NSCD is strictly a USER RESPONSIBILITY. NRCS and NCSS take no responsibility for problems that arise from use of this database.

Author(s)

The NSCD data is maintained by the NRCS NSSC Soil Survey Lab, Lincoln (NE). This subset has been prepared for the purpose of testing various DSM methods by Tomislav Hengl (tom.hengl@wur.nl). The processing steps are described in this script.

References

Soil Data Subcommittee, (1997) Soil Geographic Data Standard. Federal Geographic Data Committee, p. 286.

United States Department of Agriculture, Natural Resources Conservation Service, (2010) Keys to Soil Taxonomy, Eleventh Edition. Soil Survey Staff.

National Cooperative Soil Characterization Database (http://soils.usda.gov/survey/nscd/)

See Also

ilin.grids, geochm

Examples


## Not run: 
require(aqp)
data(ilin)
coordinates(ilin$sites) <- ~ LON + LAT
proj4string(ilin$sites) <- CRS("+proj=longlat +datum=NAD83")
# obtain state borders:
require(maps)
states.m = map('state', plot=FALSE, fill=TRUE)
IDs <- sapply(strsplit(states.m$names, ":"), function(x) x[1])
require(maptools)
states <- as(map2SpatialPolygons(states.m, IDs=IDs), "SpatialLines")
# overlay and plot points and maps:
plot(spTransform(ilin$sites, CRS("+proj=longlat +datum=WGS84")), pch="+", cex=.6, col="black")
lines(states)

## End(Not run)

[Package GSIF version 0.1-2 Index]