landmask {GSIF}	R Documentation

1-degree cell land mask for Global soil mapping purposes

Description

Land mask showing the 1-degree cells (about 19 thousand in total) in the geographical coordinates. Derived using the Global Self-consistent, Hierarchical, High-resolution Shoreline Database data (GSHHS 2.1).

Usage

data(landmask)

Format

Data frame with the following columns:

mask

boolean; land mask value

Lon_it

indication of the longitude quadrant (W or E)

Lat_it

indication of the latitude quadrant (S or N)

cell_id

cell id code e.g. W79_N83

x

longitudes of the center of the grid nodes

y

latitudes of the center of the grid nodes

Note

The land mask has been generated using the layer GSHHS_shp/h/GSHHS_h_L1.shp (level-1 boundaries). For practical purposes, all latitudes beyond 65S and 84N (polar land masses) have been masked out.

Author(s)

Tomislav Hengl

References

Wessel, P., Smith, W.H.F., (1996) A Global Self-consistent, Hierarchical, High-resolution Shoreline Database. Journal of Geophysical Research, 101, 8741-8743.

Global Self-consistent, Hierarchical, High-resolution Shoreline Database (http://en.wikipedia.org/wiki/GSHHS)

Examples


data(landmask)
gridded(landmask) <- ~x+y
proj4string(landmask) <- "+proj=longlat +datum=WGS84"
## Not run: 
spplot(landmask["mask"], col.regions="grey")

## End(Not run)

[Package GSIF version 0.1-1 Index]