landmask {GSIF}	R Documentation

1-degree cell land mask for Global soil mapping purposes

Description

Land mask showing the 1-degree cells (about 19 thousand in total) in the geographical coordinates, and the productive soils mask (areas with a positive Leaf Area Index at least once in the period 2002–2011). The land mask is based on the Global Self-consistent, Hierarchical, High-resolution Shoreline Database data (GSHHS 2.1), and the productive soils mask on the MODIS Leaf Area Index monthtly product (MOD15A2).

Usage

data(landmask)

Format

Data frame with the following columns:

mask

boolean; land mask value

soilmask

boolean; soil mask value

Lon_it

indication of the longitude quadrant (W or E)

Lat_it

indication of the latitude quadrant (S or N)

cell_id

cell id code e.g. W79_N83

x

longitudes of the center of the grid nodes

y

latitudes of the center of the grid nodes

Note

The land mask has been generated from the layer GSHHS_shp/h/GSHHS_h_L1.shp (level-1 boundaries). For practical purposes, all latitudes beyond 65S and 84N (polar land masses) have been masked out.

Author(s)

Tomislav Hengl

References

Wessel, P., Smith, W.H.F., (1996) A Global Self-consistent, Hierarchical, High-resolution Shoreline Database. Journal of Geophysical Research, 101, 8741-8743.

Savtchenko, A., D. Ouzounov, S. Ahmad, J. Acker, G. Leptoukh, J. Koziana, and D. Nickless, (2004) Terra and Aqua MODIS products available from NASA GES DAAC. Advances in Space Research 34(4), 710-714.

Global Self-consistent, Hierarchical, High-resolution Shoreline Database (http://en.wikipedia.org/wiki/GSHHS)

See Also

rworldmap::rworldmapExamples, maps::map

Examples


library(rgdal)
library(sp)

data(landmask)
gridded(landmask) <- ~x+y
proj4string(landmask) <- "+proj=longlat +datum=WGS84"
## Not run:  ## plot maps:
require(maps)
country.m = map('world', plot=FALSE, fill=TRUE)
IDs <- sapply(strsplit(country.m$names, ":"), function(x) x[1])
require(maptools)
country <- as(map2SpatialPolygons(country.m, IDs=IDs), "SpatialLines")
spplot(landmask["mask"], col.regions="grey", sp.layout=list("sp.lines", country))
spplot(landmask["soilmask"], col.regions="grey", sp.layout=list("sp.lines", country))
## also available in the Robinson projection at 20 km grid:
data(landmask20km)
image(landmask20km)

## End(Not run)

[Package GSIF version 0.3-3 Index]