riodoce {GSIF}	R Documentation

Soil samples for the Rio Doce basin

Description

Soil samples (456) collected for the Rio Doce basin in Brazil.

Usage

data(riodoce)

Format

The riodoce data set contains two data frames — sites and horizons. Sites table contains the following columns:

SOURCEID

factor; unique name that identifies a particular site

LATWGS84

numeric; latitude in decimal degrees on the NAD83 datum

LONWGS84

numeric; longitude in decimal degrees on the NAD83 datum

TAXBRC

factor; original soil class based on the Brazilian Soil Classification System

Horizons table contains the following columns:

SOURCEID

factor; unique name that identifies a particular site

HRDLOC

character; local pedological designation of the horizon e.g. "Bt1"

UHDICM

numeric; upper horizon depth from the surface in cm

LHDICM

numeric; lower horizon depth from the surface in cm

PHIHO5

numeric; pH index measured in water solution

PHIKCL

numeric; pH index measured in KCl solution

ORCDRC

numeric; weight percentage in permilles of the total organic carbon

SNDPPT

numeric; weight percentage of the silt particles (0.0002–0.05 mm)

SLTPPT

numeric; weight percentage of the sand particles (0.05–2 mm)

CLYPPT

numeric; weight percentage of the clay particles (<0.0002 mm)

BSABCL

numeric; base saturation in percent

CEXBCL

numeric; Cation Exchange Capacity in cmol(+)/kg (summary)

BLDVOL

bulk density oven dry (105 C) in g per cubic centimeter

AWAIMM

total available water capacity in mm

HORUID

factor; unique name that identifies a particular horizon

The riodoce.grids data frame contains a list of soil covariates prepared for the Rio Doce catchment in Brazil (71,492 grid nodes):

TAXBRC3

factor; dominant soil class from a 1:600,000 scale polgyon map, and based on the Brazilian Soil Classification System

GEOGLS3

factor; parent material class from the geological map

DEMSRT3

numeric; SRTM DEM

SLPSRT3

numeric; slope in degrees based on the SRTM DEM

TWISRT2

numeric; SAGA Topographic Wetness Index based on the SRTM DEM

INSSRT3

numeric; potential incoming solar radiation derived in SAGA GIS

EV1MOD3

numeric; first principal component of the monthly MODIS EVI time series data

EV2MOD3

numeric; second principal component of the monthly MODIS EVI time series data

TD1MOD3

numeric; first principal component of the 8-day MODIS day-time LST time series data

TN1MOD3

numeric; first principal component of the 8-day MODIS night-time LST time series data

x

numeric; x-coordinate in the UTM 23S coordiante system with GRS67 ellipsoid

y

numeric; y-coordinate in the UTM 23S coordiante system with GRS67 ellipsoid

Note

Soil classes (orders, sub-orders and great groups) in TAXBRC: "AR" (bare rock), "CH" (Cambissolo Húmico), "CX" (Cambissolo Háplico), "CY" (Cambissolo Flúvico), "E" (Espodossolo), "G" (Gleissolo), "LAd" (Latossolo Amarelo Distrófico), "LV" (Latossolo Vermelho), "LVA" (Latossolo Vermelho Amarelo), "M" (Chernossolo), "NV" (Nitossolo), "O" (Organossolo), "PAd" (Argissolo Amarelo Distrófico), "PVAd" (Argissolo Vermelho Amarelo Distrófico), "PVAe" (Argissolo Vermelho Amarelo Eutrófico), "PV" (Argissolo Vermelho), "PVA" (Argissolo Vermelho Amarelo), "PVd" (Argissolo Vermelho Distrófico), "PVe" (Argissolo Vermelho Distrófico), "RL" (Neossolo Litólico), "RQ" (Neossolo Quartzarenico), "RR" (Neossolo Regolítico), "RY" (Neossolo Flúvico), "T" (Luvissolo abrúptico), "V" (Vertissolo); Brazilian Soil Classification System. Soil polygon map was derived from the General soil map of Minas Gerais at scale 1:600,000. Land surface parameters: potential incoming solar radiation, SAGA Topographic Wetness Index and slope map were derived in SAGA GIS using the 1 km resolution SRTM DEM obtained from the Consultative Group on International Agricultural Research (CGIAR). Principal components for the Enhanced Vegetation Index (EVI, MOD13A3 product) and Land Surface Temperature (LST, MOD11A2 product) images were derived using the 8-day and monthly time-series MODIS images. Listed gridded layers follow a standard naming convention used by WorlGrids.org (the standard 8.3 filename convention with at most eight characters): first three letter are used for the variable type e.g. DEM (digital elevation model); the next three letters represent the data source or collection method e.g. SRT (SRTM); the 6th character is the effective scale e.g. 3 indicates the 2nd standard scale i.e. 1/120 decimal degrees (about 1 km).

Author(s)

This data set has been prepared for the purpose of testing various soil mapping methods by Eliana de Souza and Tomislav Hengl (see processing steps for more info).

References

Benedetti, M.M. and Curi, N. and Sparovek, G. and de Carvalho Filho, A. and Silva, S.H.G., (2011) Updated Brazilian's Georeferenced Soil Database – an Improvement for International Scientific Information Exchanging. In: Principles, Application and Assessment in Soil Science (Edited by E. Burcu Özkaraova Güngör), InTech, chapter 16, pp. 307-330.

EMBRAPA (2006) Sistema Brasileiro de Classificação de Solos. Centro Nacional de Pesquisa de Solos. 2nd Ed, Rio de Janeiro, Brasil.

Filho, E.F. et al. (2011) Levantamento de solos e aptidão agrícola das terras da bacia do Rio Doce, Estado de Minas Gerais. Belo Horizonte/MG, 453 p. www.meioambiente.mg.gov.br

Cooper, M., Mendes. L.M.S., Silva, W.L.C, and Sparovek, G., (2005) A National Soil Profile Database for Brazil Available to International Scientists. Soil Science Society of America Journal, 69:649-652.

Examples


library(aqp)
data(riodoce)
## Not run: # Promote to SoilProfileCollection
riodoce <- join(riodoce$horizons, riodoce$sites, type='inner')
# specify point IDs and depths:
depths(riodoce) <- SOURCEID ~ UHDICM + LHDICM
# extract site data
site(riodoce) <- ~ LONWGS84 + LATWGS84 + TAXBRC  
# generate SpatialPoints
coordinates(riodoce) <- ~ LONWGS84 + LATWGS84
proj4string(riodoce) <- "+proj=latlong +datum=WGS84"
str(riodoce)
# load the gridded data:
download.file("http://worldgrids.org/rda/riodoce.grids.rda", "riodoce.grids.rda")
load("riodoce.grids.rda")
gridded(riodoce.grids) <- ~x+y
utm.csy <- "+proj=utm +zone=23 +south +ellps=GRS67 +units=m +no_defs"
proj4string(riodoce.grids) <- CRS(utm.csy)
# plot maps:
names(riodoce.grids)
image(riodoce.grids["DEMSRT3"])
points(spTransform(riodoce@sp, CRS(utm.csy)), pch="+")
# write to a GIS format:
writeGDAL(riodoce.grids["EV1MOD3"], "EV1MOD3.tif", "GTiff", mvFlag=-99999)

## End(Not run)

[Package GSIF version 0.2-1 Index]