spc {GSIF}	R Documentation

Derive Spatial Predictive Components

Description

Derives Spatial Predictive Components for a given set of covariates. It wraps the stats::prcomp method and predicts a list principal components for an object of type "SpatialPixelsDataFrame".

Usage


## S4 method for signature 'SpatialPixelsDataFrame,formula'
spc(obj, formulaString, scale. = TRUE, 
      silent = FALSE, ...)
## S4 method for signature 'list,list'
spc(obj, formulaString, scale. = TRUE, 
      silent = FALSE, ...)

Arguments

obj
object of class "SpatialPixelsDataFrame" (must contain at least two grids) or a list of objects of type "SpatialPixelsDataFrame"

formulaString
object of class "formula" or a list of formulas

scale.
object of class "logical"; specifies whether covariates need to be scaled

silent
object of class "logical"; specifies whether to print the progress

...
additional arguments that can be passed to stats::prcomp

Value

spc returns an object of type "SpatialComponents". This is a list of grids with generic names PC1,...,PCp, where p is the total number of input grids.

Note

This method assumes that the input covariates are cross-correlated and hence their overlap can be reduced. The input variables are scaled by default and the missing values will be replaced with 0 values to reduce loss of data due to missing pixels. This operation can be time consuming for large grids.

Author(s)

Tomislav Hengl

See Also

stats::prcomp, SpatialComponents-class

Examples


# load data:
library(plotKML)
pal = rev(rainbow(65)[1:48])
data(eberg_grid)
gridded(eberg_grid) <- ~x+y
proj4string(eberg_grid) <- CRS("+init=epsg:31467")
formulaString <- ~ PRMGEO6+DEMSRT6+TWISRT6+TIRAST6
eberg_spc <- spc(eberg_grid, formulaString)
names(eberg_spc@predicted) # 11 components on the end;
## Not run: # plot maps:
rd = range(eberg_spc@predicted@data[,1], na.rm=TRUE)
sq = seq(rd[1], rd[2], length.out=48)
spplot(eberg_spc@predicted[1:4], at=sq, col.regions=pal)

## End(Not run)

[Package GSIF version 0.2-5 Index]