spfkm {GSIF}	R Documentation

Supervised fuzzy k-means on spatial pixels

Description

Runs supervised fuzzy k-means using a list of covariates layers provided as "SpatialPixelsDataFrame". If class centres and variances are not provided by the users, it first fits a multinomial logistic regression model (spmultinom) using the point data and a regression matrix, then predicts the class centres and variances based on the output from the nnet::multinom.

Usage


spfkm(formulaString, observations, covariates, ...)

Arguments

formulaString
formula string

observations
object of type "SpatialPointsData"; occurrences of factors

covariates
object of type "SpatialPixelsData" or "RasterBrick"; list of covariate layers

...
optional arguments

Details

Optional arguments include:

class.c

object of type "matrix"; class centres (see examples below)

class.sd

object of type "matrix"; class deviations (see examples below)

fuzzy.e

object of type "numeric"; fuzzy exponent

Value

Returns an object of type "SpatialMemberships" with following slots: predicted (classes predicted either by the multinomial logistic regression or fuzzy k-means), model (the multinomial logistic regression model; if available), mu (memberships derived using the fuzzy k-means), class.c (submitted or derived class centres), class.sd (submitted or derived class deviations), confusion (confusion matrix).

Note

Although nnet::multinom is consider to be robust and suited for large data sets, function might not converge in some cases or result in artifacts. If this happens try setting up the class centres and variances manually.

Author(s)

Tomislav Hengl

References

Burrough, P. A., Gaans, P.F.M., and Van Hootsmans, R., (1997) Continuous classification in soil survey: spatial correlation, confusion and boundaries. Geoderma, 77(2-4), 115–135.

Multinomial logistic regression (http://en.wikipedia.org/wiki/Multinomial_logit)

Nnet package (http://CRAN.R-project.org/package=nnet)

See Also

nnet::multinom, SpatialMemberships-class

Examples


## Not run: # load data:
library(plotKML)
data(eberg)
eberg <- eberg[runif(nrow(eberg))<.2,]
data(eberg_grid)
coordinates(eberg) <- ~X+Y
proj4string(eberg) <- CRS("+init=epsg:31467")
gridded(eberg_grid) <- ~x+y
proj4string(eberg_grid) <- CRS("+init=epsg:31467")
# derive soil predictive components:
eberg_spc <- spc(eberg_grid, ~PRMGEO6+DEMSRT6+TWISRT6+TIRAST6)
# predict memberships:
formulaString = soiltype ~ PC1+PC2+PC3+PC4+PC5+PC6+PC7+PC8+PC9+PC10
eberg_sm <- spfkm(formulaString, eberg, eberg_spc@predicted)
# plot memberships:
pal = seq(0, 1, 1/50)
spplot(eberg_sm@mu, col.regions=pal)
image(eberg_sm@mu[1], col=pal)
text(eberg@coords, paste(eberg$soiltype), cex=.6, col="black")
# classes predicted:
eberg_sm@predicted$soiltype <- as.factor(eberg_sm@predicted$soiltype)
Ls = length(levels(eberg_sm@predicted$soiltype))
pnts = list("sp.points", eberg, pch="+", cex=.6, col="black")
spplot(eberg_sm@predicted, col.regions=rainbow(Ls)[rank(runif(Ls))], sp.layout=pnts)
spplot(eberg_sm@mu[2], at=pal, col.regions=rev(heat.colors(50)), sp.layout=pnts)

## End(Not run)

[Package GSIF version 0.2-1 Index]