test.gstatModel-methods {GSIF}	R Documentation

Methods to test predictability of a regression-kriging model

Description

Test predictability of a regression-kriging model. Automates model fitting, cross-validation and prediction and prints out: (1) amount of variance explained under different sampling intensities, (2) number of predictions per second and (3) number of prediction failures (failure = predictions where cross-validation z-scores exceed +/- 1.5 or cross-validation residuals exceed three standard deviations).

Usage

 
## S4 method for signature 'SpatialPointsDataFrame,formula,SpatialPixelsDataFrame'
test.gstatModel(observations, formulaString, covariates, Ns, 
 predictionLocations = covariates, save.predictions = FALSE, ...)

Arguments

observations
object of type "SpatialPointsDataFrame" or "geosamples-class"

formulaString
object of type "formula" or a list of formulas

covariates
object of type "SpatialPixelsDataFrame", or list of grids

Ns
vector; list of sampling intensities (maximum should not exceed the total number of samples)

predictionLocations
object of class "SpatialPixelsDataFrame"; if not specified then the same as the covariates

save.predictions
logical; indicates whether the prediction results should also be saved

...
other optional arguments that can be passed to fit.gstatModel

Note

Vector of sampling intensities, if not provided, will be estimated as 10 numbers on square root scale (N minimum is determined as 20 + number of covariates times 10). Where no model can be fitted, function returns an empty set. This function can be time consuming for large data sets.

Author(s)

Tomislav Hengl, Gerard B.M. Heuvelink and Bas Kempen

See Also

fit.gstatModel, gstatModel-class

Examples


# 2D model:
library(sp)
## load the Meuse data set:
demo(meuse, echo=FALSE)
## model diagnostics:
t1 <- test.gstatModel(meuse, om~dist+ffreq, meuse.grid, family = gaussian(log), Ns=c(80, 155))
t1

[Package GSIF version 0.3-3 Index]